% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best.lambda.R
\name{best.lambda}
\alias{best.lambda}
\title{Caluculate coefficients using ridge regression and finding the best lambda for ridge regression using cross validation.}
\usage{
best.lambda(form, data, lambda)
}
\arguments{
\item{form}{dependent variable ~ all independent variables}

\item{data}{the dataframe that will be used to caluculate coefficients}

\item{lambda}{A user supplied lambda sequence. Typical usage is to have the program compute its own lambda sequence based on nlambda and lambda.min.ratio.}
}
\value{
best.lambda: best lambda
}
\description{
This is a generic function: we use ridge regression to calculate the linear coeffients between dependent variable and independent wariables and finding the best lambda for ridge regression using cross validation.
}
\examples{
best.lambda(Sepal.Length ~., iris, seq(0,1, by=0.01))
}
