% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_lambda.R
\name{best_lambda}
\alias{best_lambda}
\title{find the best lambda in ridge regression function}
\usage{
best_lambda(
  formula,
  df,
  contrasts = NULL,
  lambdas = seq(-2, 2, by = 0.01),
  nfold = 10
)
}
\arguments{
\item{formula}{an object of class "formula": describing the model to be fitted.}

\item{df}{a dataframe that containing all the variables in the model.}

\item{contrasts}{an optional list of contrasts for factor variables.}

\item{lambdas}{a sequence of lambdas to be selected from, default is seq(-1, 1, by = 0.01)}

\item{nfold}{number of folds for the cross validation}
}
\description{
This is a function taking a formula, a dataframe
and an optional list of contrast for factor variables,
and taking into account colinear (or nearly colinear) regression variables and returning coefficients.
}
\examples{
data(iris)
form <- Sepal.Length ~ Sepal.Width
best_lambda(formula = form,df = iris)

}
