% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bis557-doc.R
\docType{package}
\name{bis557}
\alias{bis557}
\alias{bis557-package}
\title{bis557 Package}
\description{
This project is created for homework 2 of BIS 557.
}
\author{
Xiaojing Zhu \email{xiaojing.zhu@yale.edu}
}
