% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bos_zip_dist}
\alias{bos_zip_dist}
\title{Final Project Dataset - Zipcode Distance from Downtown Boston}
\format{
A data frame with 30 rows and 2 variables:
\describe{
  \item{ZIPCODE}{Boston zipcodes}
  \item{DIST_FROM_DT}{Distance from zipcode to DT Boston, in miles}
}
}
\usage{
bos_zip_dist
}
\description{
This dataset has been created by myself, where I used Google Maps and its
ruler to measure the straight distance "as the crow flies" from the zipcode
to either Downtown Boston, Back Bay, or Cambridge (most expensive real
estate and most lucrative economic sectors), and choose the one with the
minimum distance. The intervals are every 2 miles.
}
\keyword{datasets}
