% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break-kkt.R
\name{break_kkt}
\alias{break_kkt}
\title{Test where the KKT Conditions are Broken}
\usage{
break_kkt(b, X_stand, y, lambda)
}
\arguments{
\item{b}{the coefficient vector to check}

\item{X_stand}{the data matrix for the LASSO regression, assumed to be standardized (mean 0 and variance 1 for each column) prior}

\item{y}{the vector of responses}

\item{lambda}{the LASSO penalty constant (we assume there is no ridge regression component in this elastic net)}
}
\value{
A logical vector of the same length as `b` that says whether each coefficient breaks the KKT Conditions
}
\description{
This function returns a logical vector of the same length as the coefficient vector that says where the Karush-Kuhn-Tucker (KKT) Conditions are broken.
}
\examples{
X <- model.matrix(Sepal.Length ~ ., iris)
plot(cv.glmnet(X, iris$Sepal.Length, standardize=TRUE)) #Get an estimate for the optimal lambda
lambda <- exp(-4.5) #Based on the above plot, after exponentiating
b <- glmnet(X, iris$Sepal.Length, lambda=lambda, standardize=TRUE)$beta
X_stand <- scale(X) #Standardize the data matrix X before plugging it into this function
break_kkt(b, X_stand, iris$Sepal.Length, lambda)
}
