% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casl_lenet.R
\name{casl_lenet}
\alias{casl_lenet}
\title{Linear Elastic Net}
\usage{
casl_lenet(
  X,
  y,
  lambda,
  alpha = 1,
  b = matrix(0, nrow = ncol(X), ncol = 1),
  tol = 1e-05,
  maxit = 50L,
  W = rep(1, length(y))/length(y)
)
}
\arguments{
\item{X}{A numeric data matrix}

\item{y}{Response vector}

\item{lambda}{The penalty term}

\item{alpha}{Value from 0 and 1; balance between l1/l2 penalty}

\item{b}{Initialize regression coefficients}

\item{tol}{Numeric tolerance parameter}

\item{maxit}{Integer maximum number of iterations}

\item{W}{Vector of sample weights}
}
\value{
Regression vector beta of length ncol(X)
}
\description{
Compute linear elastic net using coordinate descent. Need
`alpha = 1` for LASSO Regression. This function is from the BIS 557 Textbook
(Chapter 7, Page 191).
}
