% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casl_lenet_update_beta.R
\name{casl_lenet_update_beta}
\alias{casl_lenet_update_beta}
\title{Update beta vector using coordinate descent}
\usage{
casl_lenet_update_beta(X, y, lambda, alpha, b, W)
}
\arguments{
\item{X}{A numeric data matrix}

\item{y}{Response vector}

\item{lambda}{The penalty term}

\item{alpha}{Value from 0 and 1; balance between l1/l2 penalty}

\item{b}{A vector of initialized coefficients}

\item{W}{A vector of sample weights}
}
\value{
Regression vector with length ncol(X).
}
\description{
Update beta vector using coordinate descent. Need `alpha = 1`
for LASSO Regression. This function is from the BIS 557 Textbook (Chapter 7,
Page 190).
}
