% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_network.r
\name{casl_nn_forward_prop}
\alias{casl_nn_forward_prop}
\title{Apply forward propagation to a set of NN weights and biases.}
\usage{
casl_nn_forward_prop(x, weights, sigma)
}
\arguments{
\item{x}{A numeric vector representing one row of the input}

\item{weights}{A list created by casl_nn_make_weights}

\item{sigma}{The activation function.}
}
\value{
A list containing the new weighted responses (z) and activations (a).
}
\description{
Apply forward propagation to a set of NN weights and biases
}
