% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_network.r
\name{casl_nn_make_weights}
\alias{casl_nn_make_weights}
\title{Initialize list of weights to describe a dense neural network.}
\usage{
casl_nn_make_weights(sizes)
}
\arguments{
\item{sizes}{A vector giving the size of each layer, including the input and output layer}
}
\value{
A list containing initialized weights and biases.
}
\description{
Create list of weights to describe a dense neural network.
}
