% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_network.r
\name{casl_nn_sgd}
\alias{casl_nn_sgd}
\title{Apply backward propagation.}
\usage{
casl_nn_sgd(X, y, sizes, epochs, eta, weights = NULL, loss = "mse")
}
\arguments{
\item{X}{A numeric data matrix}

\item{y}{A numeric vector of responses}

\item{sizes}{A numeric vector giving the sizes of layers in the neural network}

\item{epochs}{Interger number of epochs for the algorithm to iterative for}

\item{eta}{Positive numeric learning rate}

\item{weights}{Optional list of starting weights}

\item{loss}{specifiy the loss metric to use, mean square error or mean absolute deviation}
}
\value{
A list containing the trained weights for the network
}
\description{
Apply stochastic gradient descent (SGD) to estimate NN.
}
