% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn-sgd-mae.R
\name{casl_nn_sgd_mae}
\alias{casl_nn_sgd_mae}
\title{Apply stochastic gradient descent (SGD) to estimate the weights of 
neural network with MAE as the loss function.}
\usage{
casl_nn_sgd_mae(X, y, sizes, epochs, eta, weights = NULL)
}
\arguments{
\item{X}{A numeric data matrix.}

\item{y}{A numeric vector of responses.}

\item{sizes}{A numeric vector giving the sizes of layers in
the neural network.}

\item{epochs}{Integer number of epochs to computer.}

\item{eta}{Positive numeric learning rate.}

\item{weights}{Optional list of starting weights.}
}
\description{
Estimate weights in a neural network.
}
