% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_network.r
\name{casl_util_mae_p}
\alias{casl_util_mae_p}
\title{Derivative of the mean absolute deviation (MAE) function.}
\usage{
casl_util_mae_p(y, a)
}
\arguments{
\item{y}{a numeric vector of responses}

\item{a}{a numeric vector of predicted responses}
}
\value{
the current derivative of MAE
}
\description{
Compute derivative of the mean absolute deviation (MAE) function
}
