% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_network.r
\name{casl_util_mse_p}
\alias{casl_util_mse_p}
\title{Derivative of the mean squared error (MSE) function.}
\usage{
casl_util_mse_p(y, a)
}
\arguments{
\item{y}{a numeric vector of responses}

\item{a}{a numeric vector of predicted responses}
}
\value{
the current derivative of MSE
}
\description{
Compute derivative of the mean squared error (MSE) function
}
