% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_network.r
\name{casl_util_ReLU}
\alias{casl_util_ReLU}
\title{Apply a rectified linear unit (ReLU) to a vector/matrix.}
\usage{
casl_util_ReLU(v)
}
\arguments{
\item{v}{A numeric vector or matrix}
}
\value{
The original input with negative values truncated to zero
}
\description{
Apply a rectified linear unit (ReLU) to a vector/matrix
}
