% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casl_util_soft_thresh.R
\name{casl_util_soft_thresh}
\alias{casl_util_soft_thresh}
\title{Soft threshold function}
\usage{
casl_util_soft_thresh(a, b)
}
\arguments{
\item{a}{Numeric vector of values to threshold}

\item{b}{The soft thresholded value}
}
\value{
Numeric vector of the soft-thresholded values of a.
}
\description{
Soft threshold function. This function is from the BIS 557
Textbook (Chapter 7, Page 189-190).
}
