% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_set_screening.r
\name{check_kkt}
\alias{check_kkt}
\title{Check KKT}
\usage{
check_kkt(x, y, beta, lambda)
}
\arguments{
\item{x}{design matrix}

\item{y}{response vector}

\item{beta}{regression coefficients fitted in the last iteration}

\item{lambda}{penalty parameter}
}
\value{
a logical vector indicating where KKT the condition will be violated 
for the coeffcients whose current values are 0.
}
\description{
Check KKT condition for the beta for some iteration
}
