% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_log.R
\name{class_log}
\alias{class_log}
\title{A classification model with multiple classes}
\usage{
class_log(X, y, class, niter = 100)
}
\arguments{
\item{X}{The predictors of the classification model}

\item{niter}{The number of iterations}

\item{Y}{The response variable of the classification model}
}
\value{
The estimated probabilities and the misclassification rate
}
\description{
A classification model generalizing logistic regression with multiple classes
}
\examples{
X <- matrix(rnorm(500), ncol=5)
y <- sample(c(1,2,3) , size=100, replace = T)
class_log(X,y, class=3)
}
