% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{cross_validation}
\alias{cross_validation}
\title{Cross-Validation for Ridge Regression}
\usage{
cross_validation(formula, data, folds = 5, lambdas = seq(0, 0.5, 0.01),
  contrasts = NULL)
}
\arguments{
\item{formula}{formula "y ~ ."}

\item{data}{a data frame}

\item{folds}{the number of folds used for cross-validation}

\item{lambdas}{a vector of lambdas}

\item{contrasts}{}
}
\value{
a tibble containing a summary of the statistics
}
\description{
methodology to find optimal lambdas using a given data set with ridge regression
}
