% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.gradient_descent.R
\name{cv.gradient_descent}
\alias{cv.gradient_descent}
\title{Do k-fold cross-validation for gradient descent method}
\usage{
cv.gradient_descent(formula, data, folds = 10, ...)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted. This should be a formula class argument.}

\item{data}{Specification of a dataframe that contains the variables in the model.}

\item{folds}{Number of folds used in the k-fold cross-validation}

\item{...}{other inputs that need to be passed to the function ridge_regression.}
}
\value{
a list containing the raw vector of residuals and the computed mean squared error (out of sample accuracy)
}
\description{
This function helps to conduct a k-fold cross-validation for the gradient descent method given model and data. The user need to
specified input for the function gradient_descent. More details see gradient_descent.
}
\examples{
data(iris)
cv.gradient_descent(Sepal.Length ~ ., iris)
}
