% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_ridge.R
\name{cv_ridge}
\alias{cv_ridge}
\title{Use cross validation to find the optimal ridge penalty parameter}
\usage{
cv_ridge(form, d, lambdas, tol = 1e-08, seed)
}
\arguments{
\item{form}{A formula specifying the model.}

\item{d}{A dataframe.}

\item{lambdas}{The candidate tuning parameters for the ridge penalty term}

\item{tol}{Tolerance for the condition number to truncate colinear variables}

\item{seed}{A seed to control the randomness}
}
\value{
A list with fitted coefficients, regression data X,Y and tolerance
}
\description{
Use cross validation to find the optimal ridge penalty parameter
}
\examples{
cv_ridge(Sepal.Length~.,iris,lambdas=seq(0,1,0.01),tol=1e-8,seed=1234)
}
