% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{cv_ridge_regression}
\alias{cv_ridge_regression}
\title{CV for Ridge Regression}
\usage{
cv_ridge_regression(data, formula, folds = 2, lambdas = seq(0, 0.5,
  0.01))
}
\arguments{
\item{data}{a data frame}

\item{formula}{a model formula of the form `y ~ ...`}

\item{folds}{the number of folds on which to cross validate}

\item{lambdas}{a vector of lambdas to check (default is `seq(0, 0.5, 0.01)`)}
}
\value{

}
\description{
CV for Ridge Regression
}
\examples{



}
