% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cv_test}
\alias{cv_test}
\title{Sample dataset for testing cross validation.}
\format{A data frame with 200 rows and 4 variables.
\describe{
  \item{response}{response}
  \item{pred1}{predictor 1}
  \item{pred2}{predictor 2}
  \item{pred3}{predictor 3}
  \item{pred4}{predictor 4}
}}
\usage{
cv_test
}
\description{
A dataset of modestly correlated variables from mvrnorm() and one response variable for testing cv_ridge_regression.
}
\keyword{datasets}
