% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.r
\docType{data}
\name{emnist}
\alias{emnist}
\title{emnist data for CNN modelling}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10000 rows and 785 columns.}
\source{
\url{https://www.kaggle.com/crawford/emnist#emnist-letters-train.csv}
}
\usage{
emnist
}
\description{
a 1W sample of emnist data downloaded from Kaggle
}
\keyword{datasets}
