% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_emnist_letters.R
\docType{data}
\name{emnist_letters}
\alias{emnist_letters}
\title{A data set of handwritten letters}
\description{
This data set is used to explore neural network. It contains a balanced set of the uppercase and lowercase 26 letters.
The size of training set is 124800, and the size of testing set is 20800.
}
