% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kern_density.r
\name{epan_kernel}
\alias{epan_kernel}
\title{Evaluate the Epanechnikov kernel function}
\usage{
epan_kernel(x, h = 1)
}
\arguments{
\item{x}{data upon which Epanechnikov kernel should be evaluated}

\item{h}{bandwidth for the Epanechnikov kernel}
}
\value{
a vector of values with same length as x
}
\description{
This function is used to evaluate Epanechnikov kernel function for given data.
}
