% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_propagation.R
\name{f_propagate}
\alias{f_propagate}
\title{Forward Propagation}
\usage{
f_propagate(X, W, b, sigma)
}
\arguments{
\item{X}{input}

\item{W}{weights}

\item{b}{biases}

\item{sigma}{activation function}
}
\value{
a matrix of
}
\description{
Apply forward propagation to a set of NN weights and biases.
}
\examples{
\dontrun{
f_propagate(X, W1, b1, relu)
}
}
\author{
Yijun Yang
}
