% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_clean.R
\name{final_clean}
\alias{final_clean}
\title{Final - Cleaning Boston Dataset}
\usage{
final_clean()
}
\value{
Cleaned Boston dataset
}
\description{
This function is specific to the Final dataset for Boston
properties. This cleans the raw Boston dataset so that only residential
properties are analyzed. Descriptions are inside the function code.
}
