% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{final_data_bos}
\alias{final_data_bos}
\title{Final Project Dataset - Boston 2020 Property Assessment and Taxes}
\format{
A data frame with 175072 rows and 75 variables:
\describe{
  \item{PID}{Unique 10-digit parcel number}
  \item{CM_ID}{10-digit parcel number of main condo building}
  \item{GIS_ID}{Similar to PID}
  \item{ST_NUM}{Property street number}
  \item{ST_NAME}{Property street name}
  \item{ST_NAME_SUF}{Property street name suffix}
  \item{UNIT_NUM}{Condo or apartment number}
  \item{ZIPCODE}{Property zipcode}
  \item{PTYPE}{State class code}
  \item{LU}{Type of property}
  \item{OWN_OCC}{Residential exemption of owner-occupied property}
  \item{OWNER}{Property primary owner name}
  \item{MAIL_ADDRESSEE}{Care of recipient}
  \item{MAIL_ADDRESS}{Owner Mailing address street name}
  \item{MAIL.CS}{Owner Mailing address city and state}
  \item{MAIL_ZIPCODE}{Owner Mailing address zipcode}
  \item{AV_LAND}{Total assessed land value}
  \item{AV_BLDG}{Total assessed building value}
  \item{AV_TOTAL}{Total assessed property value}
  \item{GROSS_TAX}{Property tax amount}
  \item{LAND_SF}{Parcel's land area - sq ft}
  \item{YR_BUILT}{Property - Year Built}
  \item{YR_REMOD}{Property - Year Last Remodeled}
  \item{GROSS_AREA}{Gross floor area}
  \item{LIVING_AREA}{Living area - sq ft}
  \item{NUM_FLOORS}{Property number of floors}
  \item{STRUCTURE_CLASS}{Structural classification - commercial}
  \item{R_BLDG_STYL}{Residential building style}
  \item{R_ROOF_TYP}{Residential roof type}
  \item{R_EXT_FIN}{Residential exterior finish}
  \item{R_TOTAL_RMS}{Residential total number of rooms}
  \item{R_BDRMS}{Residential number of bedrooms}
  \item{R_FULL_BTH}{Residential number of full baths}
  \item{R_HALF_BTH}{Residential number of half baths}
  \item{R_BTH_STYLE}{Residential bath 1 style}
  \item{R_BTH_STYLE2}{Residential bath 2 style}
  \item{R_BTH_STYLE3}{Residential bath 3 style}
  \item{R_KITCH}{Residential number of kitchens}
  \item{R_KITCH_STYLE}{Residential kitchen 1 style}
  \item{R_KITCH_STYLE2}{Residential kitchen 2 style}
  \item{R_KITCH_STYLE3}{Residential kitchen 3 style}
  \item{R_HEAT_TYP}{Residential heat type}
  \item{R_AC}{Residential AC existence}
  \item{R_FPLACE}{Residential number of fireplaces}
  \item{R_EXT_CND}{Residential exterior condition}
  \item{R_OVRALL_CND}{Residential overall condition}
  \item{R_INT_CND}{Residential interior condition}
  \item{R_INT_FIN}{Residential interior finish}
  \item{R_VIEW}{Residential view}
  \item{S_NUM_BLDG}{Condo main number of buildings}
  \item{S_BLDG_STYL}{Condo main building style}
  \item{S_UNIT_RES}{Condo main number of residential units}
  \item{S_UNIT_COM}{Condo main number of commercial units}
  \item{S_UNIT_RC}{Condo main number of mixed use units}
  \item{S_EXT_FIN}{Condo main exterior finish}
  \item{S_EXT_CND}{Condo main exterior condition}
  \item{U_BASE_FLOOR}{Condo unit - base floor}
  \item{U_NUM_PARK}{Condo unit - number of parking spaces}
  \item{U_CORNER}{Condo unit - corner of building? - logical}
  \item{U_ORIENT}{Condo unit - orientation within building}
  \item{U_TOT_RMS}{Condo unit - total number of rooms}
  \item{U_BDRMS}{Condo unit - number of bedrooms}
  \item{U_FULL_BTH}{Condo unit - number of full baths}
  \item{U_HALF_BTH}{Condo unit - number of half baths}
  \item{U_BTH_STYLE}{Condo unit bath 1 style}
  \item{U_BTH_STYLE2}{Condo unit bath 2 style}
  \item{U_BTH_STYLE3}{Condo unit bath 3 style}
  \item{U_KITCH_TYPE}{Condo unit type of kitchen}
  \item{U_KITCH_STYLE}{Condo unit kitchen style}
  \item{U_HEAT_TYP}{Condo unit heat type}
  \item{U_AC}{Condo unit AC existence}
  \item{U_FPLACE}{Condo unit number of fireplaces}
  \item{U_INT_FIN}{Condo unit interior finish}
  \item{U_INT_CND}{Condo unit interior condition}
  \item{U_VIEW}{Condo unit view}
}
}
\source{
\url{https://data.boston.gov/dataset/property-assessment}

\url{https://data.boston.gov/dataset/property-assessment/resource/bac18ae6-b8fd-4cd3-a61c-c5e1a11f716c}

\url{https://data.boston.gov/dataset/property-assessment/resource/8de4e3a0-c1d2-47cb-8202-98b9cbe3bd04}
}
\usage{
final_data_bos
}
\description{
This dataset is from **Analyze Boston**, a website that provides Boston's
open-source data. This dataset shows ownership information of Boston's
real estate properties (residential, commercial, industrial, etc.). For each
property, the assessed land value, the assessed building value, the
**total assessed value**, and the **annual property tax** is shown. Property
characteristics, such as the property living/lot size, year built, building
architecture style, and the total number of rooms, are shown.
}
\keyword{datasets}
