% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_kclust.R
\name{final_kclust}
\alias{final_kclust}
\title{Final - K-Means Clustering of Zipcodes}
\usage{
final_kclust(x, k)
}
\arguments{
\item{x}{The cleaned Boston dataset}

\item{k}{Number of clusters}
}
\value{
The aggregated/scaled dataset and the cluster number of each zipcode
}
\description{
This function is specific to the Final dataset for Boston
properties, which groups clusters into k clusters.
}
