% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_loss1.R
\name{final_loss1}
\alias{final_loss1}
\title{Final - Loss Function for Polynomial Model}
\usage{
final_loss1(y, y_hat, w, df)
}
\arguments{
\item{y}{Actual target vector}

\item{y_hat}{Estimated/predicted target vector (with its `names` as the test index)}

\item{w}{Vector of weights}

\item{df}{The cleaned Boston dataset}
}
\value{
The loss value
}
\description{
This function is specific to the Final dataset for Boston
properties. This is my designed loss function for the polynomial model.
}
