% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{find_lambda}
\alias{find_lambda}
\title{Lambda finder}
\arguments{
\item{formula}{a formula for a linear model to regress.}

\item{data}{the data frame we run our model on.}

\item{N}{number of folds to split the dataset}

\item{lambdas}{list of lambda you want to search from, default is from 0.1 to 0.9}
}
\description{
search the best ridge regression parameter lambda based on N folds
cross validation for out of sample accuracy.
}
\examples{
find_lambda(mpg ~.,mtcars,lambdas = c(0.1,0.2,0.5))
}
