% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_new.R
\name{gd_new}
\alias{gd_new}
\title{grad_descent() function}
\usage{
gd_new(form, data, contrasts = NULL, itr = 1e+05, L = 1e-04, tolerance = 1e-10)
}
\arguments{
\item{form}{a formula;}

\item{data}{a data frame used for the function;}

\item{contrasts}{a list of contrasts for factor variables}

\item{itr}{number of iteration}

\item{L}{learning rate for gradient descent}

\item{tolerance}{allowed tolerance for two residuals}
}
\description{
another way to build a linear model. gradient descent for ordinary least squares
}
\examples{
data(iris)
fit_linear_model <- grad_descent(Sepal.Length ~ ., iris, contrasts = list(Species = "contr.sum"))
}
