% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_outsample.R
\name{gd_outsample}
\alias{gd_outsample}
\title{gradient descent using out of sample penalty}
\usage{
gd_outsample(formula, df, contrasts = NULL, gamma = 1e-04,
  iter = 10^6)
}
\arguments{
\item{formula}{an object of class "formula": describing the model to be fitted.}

\item{df}{a dataframe that containing all the variables in the model.}

\item{contrasts}{an optional list of contrasts for factor variables.}

\item{gamma}{gamma_k, the learning rate}

\item{iter}{number of iterations}
}
\description{
This is a function taking a formula, a dataframe
and an optional list of contrast for factor variables,
and using gradient descent with out of sample penalty and returning OLS coefficients.
Large dataset should have more accurate results for splitting samples.
}
\examples{
data(iris)
form <- Sepal.Length ~ Sepal.Width
gd_outsample(formula = form,df = iris)
}
