% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_descent_glm.R
\name{glm_adapt}
\alias{glm_adapt}
\title{Gradient Descent for GLM (adaptive step)}
\usage{
glm_adapt(
  X,
  Y,
  mu_fun,
  var_fun,
  maxit = 1e+06,
  tol = 1e-10,
  gamma = 1e-05,
  m = 0.9
)
}
\arguments{
\item{X}{design matrix}

\item{Y}{response vector}

\item{mu_fun}{function from eta to the expected value}

\item{maxit}{integer maximum number of iterations}

\item{tol}{numeric tolerance parameter}

\item{gamma}{The step size}

\item{mom}{momentum parameter}
}
\value{
a list of beta coefficients
}
\description{
Implement a first-order solution for the GLM maximum likelihood problem using only gradient information, avoiding the Hessian matrix.
}
\examples{
\dontrun{
glm_adapt(X, Y, mu_fun = function(eta) 1/(1+exp(-eta)), var_fun = function(eta) eta)
}
}
\author{
Yijun Yang
}
