% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm-first-order.r
\name{glm_first_order}
\alias{glm_first_order}
\title{a first-order solution for the GLM maximum likelihood problem using only gradient information.}
\usage{
glm_first_order(X, Y, mu_fun, steps, maxit = 10000, tol = 1e-20)
}
\arguments{
\item{X}{data matrix}

\item{Y}{Response vector}

\item{mu_fun}{The inverse of link function}

\item{steps}{the step size.}

\item{maxit}{Default is 10000. The maximum number of iterations.}

\item{tol}{Default is 1e-20. The minimum difference between the old beta and the update beta.}
}
\description{
This function solves GLM with a first order solution, which avoids the Hessian matrix. Here we could use either a constant step size or a an adaptive step size.
}
\examples{
set.seed(100)
maxit=10000
steps1=rep(0.001,maxit)
n=100 
p=2
X=matrix(rnorm(n*p),n,p)
X=cbind(1,X)
beta=matrix(c(2,4,6),1,3)
prob=exp(beta\%*\%t(X))/(1+exp(beta\%*\%t(X)))
y=rbinom(n,size = 1, prob = prob)
glm_first_order(X,y, binomial(link="logit")$linkinv,steps1,maxit=maxit)
}
