% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm-firstorder.R
\name{glm_firstorder}
\alias{glm_firstorder}
\title{First-Order Method of Generalized Linear Model}
\usage{
glm_firstorder(form, data, mu_fun, step = "constant", gamma = 1e-06,
  maxit = 1000, tol = 1e-06)
}
\arguments{
\item{form}{a formula with the legal format.}

\item{data}{a dataframe provided by the user.}

\item{mu_fun}{a specified link function inverse.}

\item{step}{indicate method, constant or momentum step.}

\item{gamma}{the learning rate.}

\item{maxit}{the maximum number of iterations.}

\item{tol}{the tolerance of the squared difference of beta to show convergence.}

\item{data}{a dataframe provided by the user.}
}
\value{
a list including the coefficient estimates.
}
\description{
This function gives a first-order solution for the GLM maximum likelihood problem using only gradient information,
avoiding the Hessian matrix. A comparison of a constant step size and an adaptive one is also included.
}
\examples{
set.seed(1713948)
n <- 1000
truth <- c(4, 0.5, 0.3, 1)
X <- cbind(rep(1,n), matrix(rnorm(3*n), ncol=3))
lambda <- exp(X \%*\% truth)
Y <- rpois(n, lambda=lambda)
covs <- X[,-1]
q2.data <- data.frame(cbind(Y, covs))
form <- Y ~ .
glm_firstorder(form, q2.data, step="momentum", mu_fun=function(eta) exp(eta))$beta
}
