% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_gd.R
\name{glm_gd}
\alias{glm_gd}
\title{Implement gradient descent GLM, constant}
\usage{
glm_gd(X, Y, mu_fun, lr, max_n = 25, tol = 1e-10)
}
\arguments{
\item{X}{a model matrix, the X variables as columns}

\item{Y}{a factor vector, the response variable}

\item{mu_fun}{a function depending on Xbeta, giving the mean of the link function}

\item{lr}{the learning rate}

\item{max_n}{max number of iterations (a positive integer)}

\item{tol}{the difference threshold for which we will exit the algorithm for iterations less than max_n}
}
\description{
This function implements gradient descent GLM using both a constant adaptive step size
}
\examples{
n <- 5000; p <- 3
beta <- c(-1, 0.2, 0.1)
X <- cbind(1, matrix(rnorm(n*(p-1)), ncol = p-1))
eta <- X \%*\% beta
lambda <- exp(eta)
y <- stats::rpois(n, lambda = lambda)
beta_hat <- glm_gd(X, y, mu_fun = function(eta) exp(eta), lr=0.00000001, max_n = 100000)
}
