% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_grad.R
\name{glm_grad}
\alias{glm_grad}
\title{A first-order solution for the GLM maximum likelihood problem}
\usage{
glm_grad(
  X,
  y,
  family = poisson(link = "log"),
  niter = 1000,
  lambda = 1e-05,
  update = FALSE,
  gamma = 0.95
)
}
\arguments{
\item{X}{The predictors of the glm}

\item{y}{The response variable of the glm}

\item{family}{The link function}

\item{niter}{The number of iterations}

\item{lambda}{The learning rate}

\item{update}{Either a constant step size or an adaptive step size}

\item{gamma}{The momentum step size}
}
\value{
The estimated coefficients for the glm
}
\description{
a first-order solution for the GLM maximum likelihood problem using only gradient information without the Hessian matrix
}
\examples{
#reference CASL 5.3
n = 1000; p = 5
X <- cbind(1, matrix(rnorm(n*(p-1)), ncol=p-1))
y = sample(c(0,1), replace=TRUE, size=n)
glm(X, Y, update=FALSE)
}
