% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_gradient_adaptive.R
\name{GLM_gradient_adaptive}
\alias{GLM_gradient_adaptive}
\title{GLM_gradient_adaptive}
\usage{
GLM_gradient_adaptive(X, Y, mu_fun, mom = 0.2, maxit = 1e+05, tol = 1e-05)
}
\arguments{
\item{X}{The design matrix}

\item{Y}{Response variable}

\item{mu_fun}{A function to use eta get mu}

\item{mom}{The momentum update size, set default as 0.2}

\item{maxit}{The maximum iteration times to allow the loop to exit, set default as 1e5}

\item{tol}{The difference between previous beta and new beta to allow
the loop to exit, set default as 1e-5.}
}
\value{
A list of beta coefficients.
}
\description{
Implement a first-order solution for the GLM maximum likelihood
problem using only gradient information, avoiding the Hessian matrix.
Standard adaptive update Momentum.
}
\details{
The code is adapted from lecture notes.
}
