% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_gradient_constant.R
\name{GLM_gradient_constant}
\alias{GLM_gradient_constant}
\title{GLM_gradient_constant}
\usage{
GLM_gradient_constant(X, Y, mu_fun, maxit = 1e+05, tol = 1e-05)
}
\arguments{
\item{X}{The design matrix}

\item{Y}{Response variable}

\item{mu_fun}{A function to use eta get mu}

\item{maxit}{The maximum iteration times to allow the loop to exit, set default as 1e5}

\item{tol}{The difference between previous beta and new beta to allow
the loop to exit, set default as 1e-5.}
}
\value{
A list of beta coefficients.
}
\description{
Implement a first-order solution for the GLM maximum likelihood
problem using only gradient information, avoiding the Hessian matrix.
Constant step size.
}
\details{
The code is adapted from lecture notes.
}
