% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_hw3b.R
\name{glm_hw3b}
\alias{glm_hw3b}
\title{GLM Using First-Order Gradient Descent}
\usage{
glm_hw3b(X, y, family, steps, maxiter = 1000, tol = 1e-12)
}
\arguments{
\item{X}{Numeric (continuous or categorical) data matrix}

\item{y}{Response/Target vector}

\item{family}{Error distribution and link function, such as poisson or Gamma}

\item{steps}{Step size for each iteration}

\item{maxiter}{Maximum number of iterations}

\item{tol}{Tolerance}
}
\value{
Estimated beta coefficients
}
\description{
This function solves generalized linear models using only
first-order conditions, avoiding the Hessian matrix.
This uses the gradient descent algorithm, with either a constant or adaptive
step size.
}
\examples{
set.seed(2020)
n <- 3000; p <- 4; maxiter <- 500; steps <- rep(1e-3, maxiter)
X <- cbind(1, matrix(rnorm(n * (p-1)), ncol = p-1))
beta <- c(-1, 0.2, 0.1, 0.3)
y <- rpois(n, lambda = exp(X \%*\% beta))
glm_hw3b(X, y, family = poisson(link = "log"), steps, maxiter)

set.seed(2020)
steps2 <- 5e-3/(1:maxiter)
glm_hw3b(X, y, family = poisson(link = "log"), steps = steps2, maxiter)

}
