% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm-irwls-ridge.R
\name{glm_irwls_ridge}
\alias{glm_irwls_ridge}
\title{Create a Generalized Linear Model Using Iteratively Re-Weighted Least Squares with a Ridge Penalty}
\usage{
glm_irwls_ridge(X, y, family, lambda, maxit = 25, tol = 1e-10)
}
\arguments{
\item{X}{A numeric data matrix}

\item{y}{Response vector}

\item{family}{Instance of an R "family" object (for the family of Generalized Linear Model)}

\item{lambda}{Ridge parameter for the L-2 penalty}

\item{maxit}{Integer maximum number of iterations. Default is 25.}

\item{tol}{Numeric tolerance parameter. Default is 1e-10.}
}
\value{
Regression vector beta of length ncol(X).
}
\description{
This function is an adaptation of casl_glm_irwls from A Computational Approach to Statistical Learning to have a ridge penalty.
}
\examples{
# Based on the example in the textbook on Page 130
n <- 1000
p <- 3
X <- cbind(1, matrix(rnorm(n * (p-1)), ncol = p-1))
mu <- 1 - pcauchy(X \%*\% beta)
y <- as.numeric(runif(n) > mu)
beta <- glm_irwls_ridge(X, y, family=binomial(link="cauchit"), lambda=10)
beta
}
