% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_momentum.R
\name{glm_momentum}
\alias{glm_momentum}
\title{Implement gradient descent GLM, adaptive (momentum)}
\usage{
glm_momentum(X, Y, mu_fun, max_n, lr, gamma = 0.8, tol = 1e-10)
}
\arguments{
\item{X}{a model matrix, the X variables as columns}

\item{Y}{a factor vector, the response variable}

\item{mu_fun}{a function depending on Xbeta, giving the mean of the link function}

\item{max_n}{max number of iterations (a positive integer)}

\item{lr}{the learning rate}

\item{gamma}{the constant weight of the past update (a positive number)}

\item{tol}{the difference threshold for which we will exit the algorithm for iterations less than max_n}
}
\description{
This function implements gradient descent GLM using the momentum algorithm
}
\examples{
n <- 5000; p <- 3
beta <- c(-1, 0.2, 0.1)
X <- cbind(1, matrix(rnorm(n*(p-1)), ncol = p-1))
eta <- X \%*\% beta
lambda <- exp(eta)
y <- stats::rpois(n, lambda = lambda)
glm_momentum(X, y, mu_fun = function(eta) exp(eta), lr=0.00000001, max_n = 100000)
}
