% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_new.R
\name{glm_new}
\alias{glm_new}
\title{glm_new() function}
\usage{
glm_new(
  X,
  y,
  mu_fun,
  var_fun,
  itr = 50,
  step,
  method,
  friction,
  tolerance = 1e-10
)
}
\arguments{
\item{X}{a data matrix used for the function}

\item{y}{a vector of response values}

\item{mu_fun}{expectation function of the distribution}

\item{var_fun}{variance function of the distribution}

\item{itr}{number of iteration}

\item{step}{learning rate for gradient descent}

\item{method}{constant step size or momentum update}

\item{friction}{a number to control the velocity and prevents overshooting}

\item{tolerance}{allowed tolerance for two residuals}
}
\description{
another way to build a generalized linear model using only gradient information
}
\examples{
 n <- 5000; p <- 3; beta <- c(-1, 0.2, 0.1); X <- cbind(1, matrix(rnorm(n * (p- 1)), ncol = p - 1))
 eta <- X \%*\% beta; lambda <- exp(eta); y <- rpois(n, lambda = lambda)
   fit_linear_model <- glm_new(X, y, mu_fun = function(eta) exp(eta), var_fun = identity, step = 0.0001, method = "constant")
}
