% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW3.R
\name{glmfo}
\alias{glmfo}
\title{first-order solution for the GLM maximum likelihood}
\usage{
glmfo(x, y, family, method = "constant", lr = 1e-04, maxit = 10000,
  er = 1e-12, print = FALSE)
}
\arguments{
\item{x}{feature variables}

\item{y}{response variables}

\item{family}{family object and link function}

\item{method}{optimization method}

\item{lr}{learning rate}

\item{maxit}{Maximum number of iterations}

\item{er}{allowed errors}

\item{print}{whether to print to track progress}
}
\description{
implementation of a first-order solution for the GLM maximum likelihood problem using only gradient information.
}
\examples{
set.seed(100)
n <- 1000
X <- cbind(1, matrix(rnorm(n * 3), ncol = 3))
beta <- c(-1, 0.2, 0.5, 0.6 )
Y = rpois(n,exp(X \%*\% beta))
glmfo(X,Y,poisson(link='log'))
}
