% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMgradientMomentum.R
\name{GLMgradientMomentum}
\alias{GLMgradientMomentum}
\title{GLM Gradient descent with momentum}
\usage{
GLMgradientMomentum(
  X,
  y,
  mu_fun,
  T_fun,
  lrate = 0.01,
  mom = 0.9,
  maxiter = 10000,
  tol = 1e-05
)
}
\arguments{
\item{X}{the design matrix}

\item{y}{the response variable}

\item{mu_fun}{function from eta to the expected value}

\item{T_fun}{the sufficient statistic as a function of y}

\item{lrate}{the learning rate}

\item{mom}{the momentum parameter}

\item{maxiter}{the maximum iteration number}

\item{tol}{the numerical tolerance}
}
\description{
Fits generalized linear model by gradient descent, maximizing log-likelihood
}
