% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{grab_coeffs}
\alias{grab_coeffs}
\title{Grab the slop coefficients}
\usage{
grab_coeffs(lm_obj)
}
\arguments{
\item{lm_obj}{the linear model object created by the lm() function}
}
\description{
This is a simple unction to grab the slope coefficients
of a linear model
}
\examples{
library(palmerpenguins)
fit = lm(bill_length_mm ~ ., data = penguins[,-8])
grab_coeffs(fit)
}
