% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_coefs.R
\name{grab_coefs}
\alias{grab_coefs}
\title{Grab the slope coefficients}
\usage{
grab_coefs(lm_obj)
}
\arguments{
\item{lm_obj}{the linear model created by the lm function}
}
\description{
This function grabs the coefficients of a linear model.
}
\examples{
library(palmerpenguins)
fit <- lm (bill_length_mm ~., data = penguins[,-8])
grab_coefs(fit)
}
