% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_descent.R
\name{grad_desc}
\alias{grad_desc}
\title{Gradient Descent}
\usage{
grad_desc(formula, data, contrasts = NULL, factor = 1e-04,
  max_Iter = 5e+05, threshold = 1e-12)
}
\arguments{
\item{formula}{A formula of form "A ~ B".}

\item{data}{A dataframe.}

\item{contrasts}{A list of contrasts; default is NULL.}

\item{factor}{The factor by which to adjust beta along gradient; default is 0.0001.}

\item{max_Iter}{Maximum number of iterations before the loop exits; default is 500,000.}

\item{threshold}{Size of difference between previous and new betas at which the loop exits; default is 0.001.}
}
\value{
A list of coefficients from an lm model via gradient descent.
}
\description{
Compute a linear regression via gradient descent. Passes input to linear_model if data frame is not full rank.
}
\examples{
\dontrun{
data(iris)
grad_desc(Sepal.Length ~ ., iris)
}
}
