% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_out_of_sample.R
\name{gradient_descent_loss}
\alias{gradient_descent_loss}
\title{Loss of OLS model based on out-of-sample accuracy}
\usage{
gradient_descent_loss(form, data, alpha, num_iters, v)
}
\arguments{
\item{form}{A given formula for fitting the regression}

\item{data}{A given dataset to fit the model}

\item{alpha}{The learning rate}

\item{num_iters}{The number of iterations}

\item{v}{The number of folds when implementing n-folds cross-validation}
}
\description{
This function fits OLS model using gradient descent that calculates the
loss based on out-of-sample accuracy
}
\examples{
library(palmerpenguins)
data(penguins)
gradient_descent_loss(bill_length_mm ~ ., data = penguins[,-8],alpha=0.1,num_iters=1000)
}
