% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_descent_new.R
\name{gradient_descent_new}
\alias{gradient_descent_new}
\title{Gradient Descent with Out-of-sample Accuracy}
\usage{
gradient_descent_new(
  form,
  d,
  contrasts = NULL,
  iters = 5e+05,
  gamma = 1e-04,
  threshold = 1e-12
)
}
\arguments{
\item{form}{A formula}

\item{d}{A dataframe}

\item{contrasts}{Contrasts argument}

\item{iters}{Number of maximum iterations}

\item{gamma}{The learning rate}

\item{threshold}{Threshold for the difference between RSS of new and old betas.}
}
\value{
a list of beta coefficients
}
\description{
Implement a new function fitting the OLS model using gradient descent that calculates the penalty based on the out-of-sample accuracy.
}
\examples{
\dontrun{
gradient_descent_new(body_mass_g ~ bill_depth_mm, penguins)
}
}
\author{
Yijun Yang
}
