% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-gradient-descent.r
\name{gradient_descent_OLS_cv}
\alias{gradient_descent_OLS_cv}
\title{A function fitting the OLS model}
\usage{
gradient_descent_OLS_cv(
  formula,
  data_frame,
  nfolds = 10,
  contrasts = NULL,
  lambda = 1e-04,
  tolerence = 1e-20,
  beta1 = 1,
  max_itr = 1e+06
)
}
\arguments{
\item{formula}{a formula of linear model}

\item{data_frame}{a data_frame}

\item{nfolds}{Default is 10. the number of folds for cross validation}

\item{contrasts}{Default is NULL. a list of constasts for factor variables}

\item{lambda}{Default is 0.0001. The speed of gradient descent}

\item{tolerence}{Default is 1e-20. The minimum difference between the old ssr and the update ssr.}

\item{beta1}{Default is 1. The initial value of beta.}

\item{max_itr}{Default is 1e6. The maximum number of iterations}
}
\description{
This is a function fitting the OLS model using gradient descent that calculates the penalty based on the out-of-sample accuracy. Here we use cross validation to calculate the out-of-sample accracy.
}
\examples{
data(iris)
gradient_descent_OLS_cv(Sepal.Length ~ ., iris,nfolds=10)
}
