% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_descent_oos.R
\name{gradient_descent_oos}
\alias{gradient_descent_oos}
\title{Use gradient descent that calculates the loss
based on the out-of-sample accuracy}
\usage{
gradient_descent_oos(y, X, epsilon = 1e-04, lrate = 2, iters = 1e+05)
}
\arguments{
\item{y}{vector of the dependent variable}

\item{X}{data frame or matrix of all the independent variables (all the elements should be numetric)}

\item{epsilon}{optional; default as 0.0001, difference to determine the convergence}

\item{lrate}{learning rate}

\item{iters}{maximum number of iterations}
}
\description{
This function will apply gradient descent
to a dataset and return the coefficients and out-of-sample MSE.
This method will devide the imput dataset into two parts;
80\% as training and 20\% as test dataset. It uses the 20\% to calcualte
out of sample prediction error.
}
\examples{
data(iris)
gradient_descent_oos(iris[,1], iris[,c(2,3,4)])
}
