% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_descent_outofsample.R
\name{gradient_descent_outofsample}
\alias{gradient_descent_outofsample}
\title{Gradient descent based on out of sample accuracy}
\usage{
gradient_descent_outofsample(form, dat, gamma = 0.1, tol = 0.01, iter = 10000)
}
\arguments{
\item{form}{the input formula object}

\item{dat}{the input dataframe}

\item{gamma}{the learning rate}

\item{tol}{the numerical tolerance}

\item{iter}{the maximum iterations}
}
\description{
Fits linear model by gradient descent optimizing out of sample accuracy
}
