% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework 4.R
\name{\%*\%.sparse.matrix}
\alias{\%*\%.sparse.matrix}
\title{Define multiply method for 'sparse.matrix' class}
\usage{
\method{\%*\%}{sparse.matrix}(x, y)
}
\arguments{
\item{x}{A 'sparse.matrix' object.}

\item{y}{A 'sparse.matrix' object.}
}
\value{
A 'sparse.matrix' object, which is the result of x %*% y.
}
\description{
This function is used for defining `%*%` operation in class "sparse.matrix".
}
