% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-matrix.R
\docType{methods}
\name{\%*\%,sparse.matrix,sparse.matrix-method}
\alias{\%*\%,sparse.matrix,sparse.matrix-method}
\title{Define multiplication of sparse.matrix object.}
\usage{
\S4method{\%*\%}{sparse.matrix,sparse.matrix}(x, y)
}
\arguments{
\item{x}{a sparse.matrix object}

\item{y}{a sparse.matrix object}
}
\description{
A method for sparse.matrix multiplication.
}
