% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework 4.R
\name{\%*\%}
\alias{\%*\%}
\title{Matrix multiplication based on the class of x}
\usage{
x \%*\% ...
}
\arguments{
\item{x}{An object in some class.}
}
\value{
An object in some class.
}
\description{
This function is used for matrix multiplication.
}
