% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kern_density.r
\name{kern_density}
\alias{kern_density}
\title{Kernel Density Estimation}
\usage{
kern_density(x, x_new, h = 1)
}
\arguments{
\item{x}{training vector}

\item{x_new}{test vector}

\item{h}{bandwidth for the Epanechnikov kernel}
}
\value{
Kernel density estimation for test data.
}
\description{
This function is used for calulating the kernel density estimatior.
}
\examples{
data(ridge_train)
ridge_train_scale <- as.data.frame(scale(ridge_train))
fit <- ridge_reg(y ~. - 1, 0.01, ridge_train_scale)
summary(fit)
}
