% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_py_hw4c.R
\name{lasso_py_hw4c}
\alias{lasso_py_hw4c}
\title{OLS with LASSO Regression - Python}
\usage{
lasso_py_hw4c(y, X, lambda_val)
}
\arguments{
\item{y}{target/response variable matrix}

\item{X}{independent variables matrix (contrasts should be taken care of
here) - orthonormal matrix recommended}

\item{lambda_val}{the lambda value (regularization parameter)}
}
\value{
the best-fit coefficients with LASSO regularization
}
\description{
This function fits linear models with the LASSO regression
regularization (using the l1 norm). This method is used for subset selection.
Language is in Python 3.
}
\examples{
data(iris)
y <- matrix(iris$Sepal.Length, ncol = 1)
X <- model.matrix( ~ . - Sepal.Length - Species, data = iris)
lasso_py_hw4c(y, qr.Q(qr(X)), lambda_val = 1e-1)

}
