% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-models.r
\name{linear_model}
\alias{linear_model}
\title{Fit a linear model}
\usage{
linear_model(formula, data)
}
\arguments{
\item{formula}{formula of the regression model}

\item{data}{a data.frame}
}
\value{
An lm object
}
\description{
This function use QR decomposition to solve OLS.
}
\examples{
fit <- linear_model(Sepal.Length ~., iris)
summary(fit)
}
