% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-model-py.r
\name{linear_model_py}
\alias{linear_model_py}
\title{fit the linear model using python}
\usage{
linear_model_py(Y, X)
}
\arguments{
\item{Y}{The response variable}

\item{X}{Design matrix}
}
\description{
This a simple function to fit the linear model using python.
}
\examples{
data(iris)
Y=matrix(iris$Sepal.Length, ncol = 1)
X=model.matrix(Sepal.Length ~ .,iris)
linear_model_py(Y,X)$coefficients

}
