% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_graddescent.R
\name{lm_graddescent}
\alias{lm_graddescent}
\title{Create linear model using gradient descent}
\usage{
lm_graddescent(f, d, gamma, n, contrasts = NULL)
}
\arguments{
\item{f}{a formula the data will be fit too}

\item{d}{the data to be fit (a data frame)}

\item{gamma}{the step size (a positive number)}

\item{n}{number of iterations (a positive integer)}

\item{contrasts}{a list of variable names to be used as factor variables}
}
\description{
This function creates a linear model using gradient descent.
}
\examples{
data("iris")
irisform <- Sepal.Length ~.
lm_graddescent(irisform, iris, 0.0001, 100000)
}
