% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_graddescent_xval.R
\name{lm_graddescent_xval}
\alias{lm_graddescent_xval}
\title{Linear model using gradient descent, adjusted for out-of-sample accuracy}
\usage{
lm_graddescent_xval(f, d, gamma, n, flds)
}
\arguments{
\item{f}{a formula the data will be fit too}

\item{d}{the data to be fit (a data frame)}

\item{gamma}{the step size (a positive number)}

\item{n}{number of iterations (a positive integer)}

\item{flds}{a list of folds}
}
\description{
This function creates a linear model using gradient descent, adjusting the penalty for out-of-sample accuracy.
}
\examples{
data("iris")
irisform <- Sepal.Length ~.
irisfold <- rsample::vfold_cv(iris)
lm_graddescent_xval(irisform, iris, 0.0001, 100000, irisfold)
}
