% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm-lasso-py.r
\name{lm_lasso_py}
\alias{lm_lasso_py}
\title{LASSO regression using python}
\usage{
lm_lasso_py(Y, X, lambda)
}
\arguments{
\item{Y}{The response of the data}

\item{X}{Design matrix}

\item{lambda}{The penalty parameter for LASSP regression}
}
\description{
This function is for lasso regression using python
}
\examples{
data(iris)
Y=matrix(iris$Sepal.Length, ncol = 1)
X=model.matrix(Sepal.Length ~ .,iris)
lm_lasso_py(Y,X,lambda=0)$coefficients
}
