% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm-lasso-py-or.r
\name{lm_lasso_py_or}
\alias{lm_lasso_py_or}
\title{Lasso regression using python when the design matrix is orthogonal}
\usage{
lm_lasso_py_or(Y, X, lambda)
}
\arguments{
\item{Y}{The response of the data}

\item{X}{Design matrix}

\item{lambda}{The penalty parameter for ridge regression}
}
\description{
This function is for lasso regression using python
}
\examples{
n=1000
p=4
X <- matrix(rnorm(n*p),n,p)
beta_true <- matrix(c(2,4,1.5,5),p,1)
Y <- X\%*\%beta_true
lm_lasso_py_or(Y,X,lambda=0.01)$coefficients
}
