% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_ooc.R
\name{lm_ooc}
\alias{lm_ooc}
\title{Out-of-core LM implementation}
\usage{
lm_ooc(form, d_name, nrows)
}
\arguments{
\item{form}{a formula}

\item{d_name}{the name of the data to be fit (a CSV file)}

\item{nrows}{the chunk size}
}
\description{
This function uses the Python script lm_calc to fit a linear model using small chunks of observations, rather than the whole data set at once.
}
\examples{
data("iris")
library(reticulate)
reticulate::source_python('../python/Scripts/lm_calc.py')
reticulate::use_condaenv("r-reticulate")
data(iris)
write.table(x = iris, file = "iris.csv", sep = ",", row.names = FALSE)
irisform <- Sepal.Length ~ Sepal.Width + Petal.Length + Petal.Width
beta_hats <- bis557::lm_ooc(irisform, "iris.csv", nrows=75)
print(beta_hats)
}
