% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_patho-doc.R
\docType{data}
\name{lm_patho}
\alias{lm_patho}
\title{Example data for testing use}
\description{
This is a very small dataset with 2 observations and 3 variables. The source of the dataset is the supplementry materials distributed along with Homework 1.
This dataset is mainly used for testing the functions in this package.
}
\keyword{data}
