% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm-ridge.r
\name{lm_ridge}
\alias{lm_ridge}
\title{Ridge regression}
\usage{
lm_ridge(formula, data_frame, contrasts = NULL, lambda)
}
\arguments{
\item{formula}{a formula of linear model}

\item{data_frame}{a data_frame}

\item{contrasts}{Default is NULL. a list of constasts for factor variable}

\item{lambda}{The penalty parametor for ridge regression}
}
\description{
This function is for ridge regression
}
\examples{
data(iris)
lm_ridge(Sepal.Length ~ ., iris,nfolds=10)$coefficients
}
