% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm-ridge-py.r
\name{lm_ridge_py}
\alias{lm_ridge_py}
\title{Ridge regression using python}
\usage{
lm_ridge_py(Y, X, lambda)
}
\arguments{
\item{Y}{The response of the data}

\item{X}{Design matrix}

\item{lambda}{The penalty parameter for ridge regression}
}
\description{
This function is for ridge regression using python
}
\examples{
data(iris)
Y=matrix(iris$Sepal.Length, ncol = 1)
X=model.matrix(Sepal.Length ~ .,iris)
lm_ridge_py(Y,X,lambda=0.5)$coefficients
}
