% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression_multiclasses.R
\name{logistic_regression_multiclasses}
\alias{logistic_regression_multiclasses}
\title{logistic_regression_multiclasses}
\usage{
logistic_regression_multiclasses(X, Y, mu_fun, maxit = 1e+05, tol = 1e-05)
}
\arguments{
\item{X}{The design matrix}

\item{Y}{Response variable}

\item{mu_fun}{A function to use eta get mu}

\item{maxit}{The maximum iteration times to allow the loop to exit, set default as 1e5}

\item{tol}{The difference between previous beta and new beta to allow
the loop to exit, set default as 1e-5.}
}
\value{
A list of beta coefficients.
}
\description{
Implement a classification model generalizing
logistic regression to accommodate more than two classes.
}
\details{
The code is adapted from lecture notes.
}
