% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_multiclass.R
\name{logit_multiclass}
\alias{logit_multiclass}
\title{A logistic model works for more than 2 classes}
\usage{
logit_multiclass(X, y, iters = 25, epsilon = 1e-10)
}
\arguments{
\item{X}{data frame or matrix of all the independent variables (all the elements should be numetric)}

\item{y}{vector of the dependent variable}

\item{iters}{maximum number of iterations}

\item{epsilon}{error tolerence}
}
\description{
This function implement a logistic model which can make prediction of more than 2 classes
}
\examples{
# Momentum Algorithm
X <- cbind(rep(1,100),matrix(rnorm(1000),200))
##poisson silulation
beta <- c(1, 0.1, 0.2, 0.1, 0.3, -1)
y <- rpois(nrow(X), exp(X\%*\%beta))
data <- as.data.frame(cbind(y,X[,-1]))
# with adaptive step size update
gradient_descent_mmt(y,X,family = poisson(link = "log"),update = T, lrate)
}
