% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_model_matrices.R
\name{make_model_matrices}
\alias{make_model_matrices}
\title{Make Model Matrices}
\usage{
make_model_matrices(form, d, contrasts = NULL)
}
\arguments{
\item{form}{a formula}

\item{d}{a dataframe}

\item{contrasts}{contrasts argument}
}
\value{
X and Y for the model
}
\description{
Make a model matrix easily.
}
\author{
Yijun Yang
}
