% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_model_matrix.R
\name{make_model_matrix}
\alias{make_model_matrix}
\title{Produce a list of outcome as well as model matrix}
\usage{
make_model_matrix(form, data)
}
\arguments{
\item{form}{A formula that denotes the model of interest}

\item{data}{A dataset that contains the variables of interest}
}
\description{
This function helps to generate the outcome vector and the model matrix based on the model and data given
}
\examples{
  data("iris")
  form <- Sepal.Length ~.
make_model_matrices(form, iris)
}
