% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_weights.R
\name{make_weights}
\alias{make_weights}
\title{Create Weights for Neural Networks}
\usage{
make_weights(sizes)
}
\arguments{
\item{sizes}{a vector giving the size of each layer, including the input and output layers}
}
\value{
a list containing initialized weights and biases
}
\description{
Create list of weights and biases to describe a dense neural network.
}
\examples{
\dontrun{
make_weights(c(10, 10, 10, 5))
}
}
\author{
(modify based on CASL p217)
}
