% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-matrices.R
\name{model_matrices}
\alias{model_matrices}
\title{Extract X and Y from data}
\usage{
model_matrices(form, d)
}
\arguments{
\item{form}{a formula with the legal format.}

\item{d}{a dataframe provided by the user.}
}
\description{
Make model matrices
}
\examples{
data(iris)
fit <- model_matrices(Sepal.Length ~ ., iris)
fit$X
}
