% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_logistic.R
\name{multi_logistic}
\alias{multi_logistic}
\title{Logistic regression for K classes}
\usage{
multi_logistic(X, y, maxit = 100)
}
\arguments{
\item{X}{A matrix of independent variables}

\item{y}{A vector of response variable}

\item{maxit}{Maximum number of iterations}
}
\value{
A list of fitted probabilities, classes and misclassification rate.
}
\description{
Logistic regression for K classes
}
\examples{
\dontrun{multi_logistic(X=iris[,-5],y=iris[,5],maxit=20)}
}
