% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_logreg.R
\name{multi_logreg}
\alias{multi_logreg}
\title{multi-class logistic regression}
\usage{
multi_logreg(X, y, iter = 35, tol = 1e-10)
}
\arguments{
\item{X}{numeric data matrix}

\item{y}{response vector}

\item{iter}{max number of iterations, default is 35}

\item{tol}{tolerance, default is 1e-10}
}
\description{
This is a function taking X, y and fit a multi-class logistic regression
with reference to textbook p130
}
\examples{
X <- matrix(c(1,2,3,4,5), nrow = 5)
y <- c(0,0,1,1,2)
multi_logreg(X,y)

}
