% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW3.R
\name{multiclass}
\alias{multiclass}
\title{Multiclass Logistic regression}
\usage{
multiclass(formula, data, maxit = 50, er = 1e-04)
}
\arguments{
\item{formula}{feature variables and response variables specification}

\item{data}{dataframe}

\item{maxit}{maximum iteration}

\item{er}{allowed errors}
}
\description{
implementation of multiclass logistic regression capable of classifying more than 2 classes
}
\examples{
multiclass(Species ~.,iris)
}
