% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass_hw3c.R
\name{multiclass_hw3c}
\alias{multiclass_hw3c}
\title{Multi-Class Logistic Regression}
\usage{
multiclass_hw3c(X, y, maxiter = 50, tol = 1e-12)
}
\arguments{
\item{X}{Numeric (continuous or categorical) data matrix}

\item{y}{Categorical Response/Target vector (should be a factor)}

\item{maxiter}{Maximum number of iterations}

\item{tol}{Tolerance}
}
\value{
The predicted class/category for each observation
}
\description{
This function classifies observations for multiple classes.
This algorithm will use the "one-vs-all" approach.
}
\examples{
data(iris)
multiclass_hw3c(X = iris[,-5], y = iris$Species, maxiter = 60)

}
