% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_gd.R
\name{my_gd}
\alias{my_gd}
\title{This function will fit a linear model using gradient descent.}
\usage{
my_gd(form, d, eta = 0.05, iters = 1e+05, contrasts = NULL)
}
\arguments{
\item{form}{A formula specifying the model.}

\item{d}{A dataframe.}

\item{eta}{Learning rate. Default is 0.05.}

\item{iters}{Total iterations.}

\item{contrasts}{A list specifying the contrasts. The default is NULL.}
}
\value{
beta Fitted coefficients.
}
\description{
This function will fit a linear model using gradient descent.
}
\examples{
my_gd(Sepal.Length~.,iris)
}
