% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM.R
\name{my_glm}
\alias{my_glm}
\title{First-Order GLM (Using Gradient Descent)}
\usage{
my_glm(form, data, mu_fun, method, gamma = 1e-04, maxit = 1000, tol = 1e-06)
}
\arguments{
\item{form}{A prespecified formula of interest.}

\item{data}{A dataset contains the elements of interests.}

\item{mu_fun}{The inverse link function specified to decode the linear outcome eta to the mean.}

\item{method}{Indicator of method used. If method=1, do the constant step size iteration. If method=2, do the adaptive (Momentum) one.}

\item{maxit}{Max times of iteration}

\item{tol}{A tolerance that helps to contol the cease of the iterations}

\item{Gamma}{A learning rate.}
}
\value{
An object contains the fitted coefficients and the difference that measures convergence.
}
\description{
This function provides two ways to imitate the "glm" function, solving the estimated coefficients in a first-order gradient
descient feature. The first method implements the straight forward constant learning rate one. The second method implements the adaptive
or called the Momentum algorithm of optimzation. An object contains the estimates as well as the tracking vector of difference of estimates
along the iterations would be the output.
}
\examples{
\dontrun{my_glm(form, data, mu_fun = function(eta) 1/(1+exp(-eta)), method=1)}
}
