% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm_gd_out_of_sample.R
\name{my_lm_gd_out_of_sample}
\alias{my_lm_gd_out_of_sample}
\title{OLS using gradient descent and compute out-of-sample error}
\usage{
my_lm_gd_out_of_sample(form, d, eta = 0.05, iters = 1e+05, seed = 1234)
}
\arguments{
\item{form}{A formula specifying the model.}

\item{d}{A dataframe.}

\item{eta}{Learning rate. Default is 0.05.}

\item{iters}{Total iterations.}

\item{seed}{A number controlling the random sampling.}
}
\value{
A list with fitted coefficients and out-of-sample mean squared error
}
\description{
OLS using gradient descent and compute out-of-sample error
}
\examples{
my_lm_gd_out_of_sample(Sepal.Length~.,iris,eta=0.05,iters=100000,seed=1234)
}
