% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_multinom.R
\name{my_multinom}
\alias{my_multinom}
\title{Implement multinomial logistic regression}
\usage{
my_multinom(X, Y, lr, max_n, tol = 1e-10)
}
\arguments{
\item{X}{a model matrix, the X variables as columns}

\item{Y}{a factor vector, the response variable}

\item{lr}{the learning rate}

\item{max_n}{max number of iterations (a positive integer)}

\item{tol}{the difference threshold for which we will exit the algorithm for iterations less than max_n}
}
\description{
This function implements multinomial logistic regression, generalizing to k classes
}
