% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ridge.R
\name{my_ridge}
\alias{my_ridge}
\title{Fitting a linear model using ridge regression}
\usage{
my_ridge(form, d, lambda, tol = 1e-08)
}
\arguments{
\item{form}{A formula specifying the model.}

\item{d}{A dataframe.}

\item{lambda}{The tuning parameter for the ridge penalty term}

\item{tol}{Tolerance for the condition number to truncate colinear variables}
}
\value{
A list with fitted coefficients, regression data X,Y and tolerance
}
\description{
Fitting a linear model using ridge regression
}
\examples{
my_ridge(Sepal.Length~.,iris,lambda=0.01,tol=1e-8)
}
