% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ridge_xval.R
\name{my_ridge_xval}
\alias{my_ridge_xval}
\title{Finding optimal lambda using out of sample cross validation}
\usage{
my_ridge_xval(f, d, lambdas, flds)
}
\arguments{
\item{f}{a formula the data will be fit to}

\item{d}{the data to be fit (a data frame)}

\item{lambdas}{a vector of candidate ridge parameters}

\item{flds}{a list of folds}
}
\description{
This function cross-validates a ridge regression model for out of sample accuracy for several lambdas, then finds the optimal lambda. We choose lambda that yields the lowest mean squared difference in out-of-sample residuals
}
\examples{
data("iris")
irisform <- Sepal.Length ~.
irisfold <- rsample::vfold_cv(iris)
lambda_vec <- seq(1,100,by=0.1)
my_ridge_xval(irisform, iris, lambda_vec, irisfold)
}
