% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_python.R
\name{ols_python}
\alias{ols_python}
\title{Fitting a linear model using python in R}
\usage{
ols_python(X, Y)
}
\arguments{
\item{X}{A design matrix.}

\item{Y}{A response vector.}
}
\description{
Fitting a linear model using python in R
}
\examples{
\dontrun{ols_python(X,Y)}
}
