% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeregression.R
\name{opt_lambda}
\alias{opt_lambda}
\title{opt_lambda}
\usage{
opt_lambda(lambdas, data, form)
}
\arguments{
\item{lambdas}{A numeric vector containing all the candidates for lambda.}

\item{data}{The data set to do ridge regression with.}

\item{form}{A regression formula you want to use.}
}
\value{
A number indicating the optimal lambda.
}
\description{
This is a function called opt_lambda,
which optimize the ridge regression parameter lambda by spliting data into test and trian subset
and estamate the rmse for test data based on the ridge regression results on train data with selected lambda.
The lambda with least rmse is the optimal.
}
\examples{
opt_lambda(lambdas=seq(0, 0.9, 0.005),iris,Sepal.Length~.)
}
