% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_ridge_lambda.R
\name{optim_lamb}
\alias{optim_lamb}
\title{optimize a lambda for a ridge regression model by cross validation}
\usage{
optim_lamb(X, y, lambdas, folds)
}
\arguments{
\item{X}{The predictors of the ridge regression model}

\item{Y}{The response variable of the ridge regression model}

\item{lambda}{possible penalizing coefficients}
}
\value{
The optimized lambda
}
\description{
A function that chooses the best lambda for a ridge regression
}
\examples{
 X = matrix(c(rep(1, 10), seq(2,20,2)+rnorm(10)), ncol=2)
 y = c(seq(10,100,10)+rnorm(10))
 optim_lamb(X,y, lambdas=c(1,2,3,4,5), 5)
}
