% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_lambda.R
\name{optimal_lambda}
\alias{optimal_lambda}
\title{Selection of optimal penalty strength Ridge regression model}
\usage{
optimal_lambda(X, y, lambdas, n)
}
\arguments{
\item{X}{the data design matrix}

\item{y}{the response vector}

\item{lambdas}{an array of penalty strengths}

\item{n}{number of folds in cross validation}
}
\description{
Selection of optimal penalty strength for Ridge regression model using n-fold cross validation
}
