% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{optimization}
\alias{optimization}
\title{optimization() function}
\usage{
optimization(
  form,
  dataset,
  contrasts = NULL,
  lambdas = seq(-2, 2, 0.01),
  nfold = 10
)
}
\arguments{
\item{form}{a formula;}

\item{dataset}{a data frame used for the function;}

\item{contrasts}{a list of contrasts for factor variables}

\item{lambdas}{a sequence for selection}

\item{nfold}{number of folds to be used}
}
\description{
one way to select the best lambda forridgr regressions.
}
\examples{
data(iris)
fit_linear_model <- optimization(Sepal.Length ~ ., iris, contrasts = list(Species = "contr.sum"), lambda = seq(0, 2, 0.01))
}
