% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_lambda.R
\name{optimization_lambda}
\alias{optimization_lambda}
\title{optimization_lambda}
\usage{
optimization_lambda(
  form,
  dat,
  folds = 10,
  lambdas = seq(0, 1, 0.1),
  contrasts = NULL
)
}
\arguments{
\item{form}{A formula with the format of "Y ~ .".}

\item{dat}{A dataframe.}

\item{folds}{The number of folds to cross validate}

\item{lambdas}{A list of the ridge penalty term lambda.}

\item{contrasts}{A list of contrasts.}
}
\value{
The ridge regression parameter lambda that minimizes mse.
}
\description{
Optimizing the ridge parameter lambda by cross validation
}
