% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{optimize}
\alias{optimize}
\title{lambda optimization using k-fold cross validation.}
\usage{
optimize(X, Y, k, start, end, length)
}
\arguments{
\item{X}{X from input dataset.}

\item{Y}{Y from Input dataset.}

\item{k}{number of folds for the cross validation}

\item{start}{start of test lambda sequence}

\item{end}{end of test lambda sequence}

\item{length}{number of items in the lambda sequence}
}
\value{
The estimated model as a list.
}
\description{
lambda optimization using k-fold cross validation.
}
