% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_lambda.R
\name{optimize_lambda}
\alias{optimize_lambda}
\title{Optimize Lambda for a given ridge regression function}
\usage{
optimize_lambda(formula, data, lambda.set, folds = 10, ...)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted. This should be a formula class argument.}

\item{data}{Specification of a dataframe that contains the variables in the model.}

\item{lambda.set}{a set of lambdas to be checked. This should be a vector of values or a sequence of numbers.}

\item{folds}{Number of folds used in the k-fold cross-validation}

\item{...}{other inputs that need to be passed to the function ridge_regression.}
}
\value{
a optimal value from the specified set of lambda which returns the smallest MSE in the cross-validation check, the vector of MSE over all lambdas,
and the minimum of the MSE.
}
\description{
This function help to choose optimal lambda parameter from a given set by conducting cross-validation to the models with different
lambdas and return the lambda that gives the smallest mean squared error. This function was based on the function ridge_regression.
}
\examples{
data(iris)
optimize_lambda(Sepal.Length ~ ., iris, seq(0,2,0.01))
}
