% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_ridge.R
\name{optimize_ridge}
\alias{optimize_ridge}
\title{Optimizing the ridge parameter}
\usage{
optimize_ridge(X, Y, lambdas, k)
}
\arguments{
\item{X}{The input design matrix.}

\item{Y}{The input response vector.}

\item{lambdas}{The list of input penalty parameters.}

\item{k}{An optional number k for k-fold cross-validation.}
}
\description{
This is a function to optimize the ridge penalty parameter.
}
\examples{
result <- optimize_ridge(X, Y, lambda=seq(0.1,15,length.out=50),k=5)
print(result)
}
