% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimized-lambda.R
\name{optimized_lambda}
\alias{optimized_lambda}
\title{Optimizing ridge parameter based on MSE}
\usage{
optimized_lambda(form, data, lambdas, fold.num = 10)
}
\arguments{
\item{form}{a formula with the legal format.}

\item{data}{a dataframe provided by the user.}

\item{lambdas}{a set of ridge parameter from which the function searches the best one.}

\item{fold.num}{number of folds divided for the cross validation (default=10).}
}
\value{
a list including the MSE vector, the optimized ridge parameter, and the minimimal MSE which indicates the best lambda.
}
\description{
This is a function that attempts the best lambda parameter.
}
\examples{
data(iris)
best.lambda <- optimized_lambda(Sepal.Length ~ ., iris, lambdas=seq(0, 2, by=0.01))
}
