% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizing_lambda.R
\name{optimizing_lambda}
\alias{optimizing_lambda}
\title{optimizing_lambda}
\usage{
optimizing_lambda(form, data, folds = 10, lambdas = seq(0, 1, 0.05),
  contrasts = NULL)
}
\arguments{
\item{form}{A symbolic description of the model to be fitted.}

\item{data}{A data frame or list that contains the variables in the model.}

\item{folds}{number of subsets of the data that is used in training and testing sample}

\item{lambdas}{Penalty added to beta}

\item{contrasts}{a list of contrasts}
}
\value{
The function returns a tibble that summarizes rmse and lambda
}
\description{
optimizing_lambda is used to find the optimal lambda in ridge regression by using cross validation
}
