% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_gradient_descent.R
\name{osm_gradient_descent}
\alias{osm_gradient_descent}
\title{out-of-sample gradient descent}
\usage{
osm_gradient_descent(
  form,
  dat,
  contrasts = NULL,
  factor = 1e-04,
  m.iteration = 1e+05,
  b.difference = 1e-10,
  F = 10
)
}
\arguments{
\item{form}{A formula with the format of "Y ~ .".}

\item{dat}{A dataframe.}

\item{contrasts}{A list of contrasts.}

\item{factor}{The factor to adjust beta, set default as 0.0001.}

\item{m.iteration}{The maximum iteration times to allow the loop to exit, set default as 1e5}

\item{b.difference}{The difference between previous sum of residuals and new sum of residuals
to allow the loop to exit, set default as 1e-10.}

\item{F}{The number of folds, set default as 10}
}
\value{
A list of coefficients from an linear regression model with gradient descent.
}
\description{
Implement gradient descent for ordinary least squares
}
\examples{
\dontrun{
osm_gradient_descent(Sepal.Length ~ ., iris, contrasts)
}

}
