% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{out_gradient_descent}
\alias{out_gradient_descent}
\title{out-of-sample gradient descent}
\usage{
out_gradient_descent(formula, data, lambda = 1e-04, N = 5e+05,
  M = 10)
}
\arguments{
\item{formula}{a formula for ridge regression}

\item{data}{the data you want to regress on}

\item{lambda}{learning rate}

\item{N}{number of iterations}

\item{M}{number of folds to split the dataset}
}
\description{
implementation of gradient descent for ordinary least squares, finding the solutions
for Y = X*beta and choose the best best beta that gives best out of sample accuracy
}
\examples{
set.seed(100)
ogd = data.frame(red = runif(100,0,1), yellow = runif(100,0,1),blue=runif(100,0,1))
ogd$price = 0.2*ogd$red + (0.4*(ogd$yellow **2)) + 0.5 * ogd$blue
formula = price ~ .
data = ogd
out_gradient_descent(formula, data)
}
