% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penguinsi-doc.R
\docType{data}
\name{penguinsi}
\alias{penguinsi}
\title{Date from Penguins, with imputed values}
\description{
The penguins data set after imputing all the missing values using the {missForest} package.
Note: Even imputed outcomes.
}
\references{
See the palmerpenguins::penguins documentation for information about the values.
}
\keyword{data}
