% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boundary.R
\name{plot_boundary}
\alias{plot_boundary}
\title{Plot Neural Network Resulting Classifier}
\usage{
plot_boundary(X, y, model, topic)
}
\arguments{
\item{X}{input X}

\item{y}{input y}

\item{model}{weights and biases from neural_network function}

\item{title}{title of the plot}
}
\value{
a plot with classification
}
\description{
plot the results of decision boundary of an artificial neural network
}
\examples{
\dontrun{
plot_boundary(X, y, nn.model)
}
}
\author{
Yijun Yang
}
