% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework-4.R
\name{+.sparse.matrix}
\alias{+.sparse.matrix}
\title{Define add method for 'sparse.matrix' class}
\usage{
\method{+}{sparse.matrix}(x, y)
}
\arguments{
\item{x}{A 'sparse.matrix' object.}

\item{y}{A 'sparse.matirx' object.}
}
\value{
A 'sparse.matrix' object, which is the result of x + y.
}
\description{
This function is used for defining `+` operation in class "sparse.matrix".
}
\examples{
sm2 <- sparse.matrix(i = c(1, 2, 2), j = c(1, 1, 2), x = c(4.4, 1.2, 3), dims = c(2, 3))
sm3 <- sparse.matrix(i = rep(1, 3), j = 1:3, x = 1:3, dims = c(2, 3))
sm2 + sm3
}
