% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.Gradient_Descent.R
\name{predict.gd_object}
\alias{predict.gd_object}
\title{Prediction based on a gradient descent object}
\usage{
\method{predict}{gd_object}(object, ...)
}
\arguments{
\item{object}{A gradient descent model object}

\item{...}{An input dataset which we need prediction of.}
}
\description{
This function takes in a gradient descent model object and a data set, returns the
prediction based on the model.
}
\examples{
library(palmerpenguins)
data(penguins)
my_fit =gradient_descent(bill_length_mm ~ ., data = penguins[,-8],alpha=0.1,num_iters=1000)
my_gd_predict = predict.gd_object(my_fit,data = penguins[,-8])
}
