% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.gradient_descent.R
\name{predict.gradient_descent}
\alias{predict.gradient_descent}
\title{Formulate prediction function for gradient descent}
\usage{
\method{predict}{gradient_descent}(object, ...)
}
\arguments{
\item{object}{ridge_regression object, see ridge_regression}

\item{...}{input of dataframe}
}
\value{
Predicted value of variable of interest given X's.
}
\description{
Formulate prediction function for gradient descent
}
