% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge-prediction.r
\name{predict.ridge_reg}
\alias{predict.ridge_reg}
\title{Predict the Response to a Ridge Regression}
\usage{
\method{predict}{ridge_reg}(object, ...)
}
\arguments{
\item{object}{a ridge_reg object}

\item{...}{all the other possible parameters for the predict method}
}
\value{
A vector of the predicted values for the given data and ridge regression coefficients
}
\description{
This method predicts the response to a ridge regression object.
}
\examples{
fit_reg <- ridge_reg(Sepal.Length ~., 1, iris)
predict(fit_reg, iris)
}
