% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ridge_regression.R
\name{predict.ridge_regression}
\alias{predict.ridge_regression}
\title{Formulate prediction function for Ridge Regression}
\usage{
\method{predict}{ridge_regression}(object, ...)
}
\arguments{
\item{object}{ridge_regression object, see ridge_regression}

\item{...}{input of dataframe}
}
\value{
Predicted value of variable of interest given X's.
}
\description{
Formulate prediction function for Ridge Regression
}
