% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_lm_graddescent.R
\name{predict_lm_graddescent}
\alias{predict_lm_graddescent}
\title{Prediction function for class lm_graddescent}
\usage{
predict_lm_graddescent(object, ...)
}
\arguments{
\item{object}{the lm_graddescent object}

\item{...}{the data frame used to fit the object}
}
\description{
This function predicts a value of Y based off a graddescent linear model
}
\examples{
data("iris")
irisdf <- as.data.frame(iris)
irissamdf <- irisdf[sample(nrow(irisdf), 0.2*nrow(irisdf)), ]
irisform <- Sepal.Length ~.
obj <- lm_graddescent(irisform, iris, 0.0001, 100000)
predict_lm_graddescent(obj, irisdf)
}
