% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{predict_logistic}
\alias{predict_logistic}
\title{Multiclass logistic regression prediction}
\usage{
predict_logistic(X, Y, object)
}
\arguments{
\item{X}{A specifeid design matrix}

\item{Y}{The multilevel outcome of interest. This is the reference value to evaluate the accuracy of prediction.}

\item{object}{Output from the multiclass_logistic regression function. (beta matrix)}
}
\value{
A series outcome of prediction including a matrix of conditional probabilities and a crude prediction of Y.
}
\description{
This function use the outcome of multilevel logistic regression function to conduct prediction given data.
}
