% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ridge.R
\name{predict_ridge}
\alias{predict_ridge}
\title{Predict function for ridge regression}
\usage{
predict_ridge(object, ...)
}
\arguments{
\item{object}{object of class inheriting from "ridge_regression"}

\item{...}{other arguments}
}
\description{
Predicted values based on the linear model object
}
\examples{
data(iris)
fit_ridge <- ridge_regression(form = Sepal.Length ~ ., data = iris, lambda = 0.001)
predict_ridge(fit_ridge)
}
