% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.r
\docType{data}
\name{predict.test}
\alias{predict.test}
\alias{predict_test}
\title{A data set for the 5th homework}
\description{
In homework 5, this is the predicted value for problem2.

In homework 5, this is the predicted value for problem2.
}
