% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.r
\docType{data}
\name{predict.test.1}
\alias{predict.test.1}
\alias{predict_test_1}
\title{A data set for the 5th homework}
\description{
In homework 5, this is the predicted value for problem1.

In homework 5, this is the predicted value for problem1.
}
