% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.r
\docType{data}
\name{predict_train}
\alias{predict_train}
\alias{predict.train}
\title{A data set for the 5th homework}
\description{
In homework 5, this is the predicted value for problem2.

In homework 5, this is the predicted value for problem2.
}
