% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pylinear.R
\name{pylinear}
\alias{pylinear}
\title{Fitting a linear model using Python.}
\usage{
pylinear(X, y)
}
\arguments{
\item{X}{The input design matrix.}

\item{y}{The input response vector.}
}
\value{
A list of estimated coefficients.
}
\description{
This is a function fits linear model.
}
\examples{
\dontrun{pylinear(X,y)}
}
