% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relu.R
\name{relu}
\alias{relu}
\title{A Rectified Linear Unit (ReLU)}
\usage{
relu(v)
}
\arguments{
\item{v}{a numeric vector or matrix}
}
\value{
the original vector with negative values truncated to zero
}
\description{
pushes negative values to 0 while returns positive values unmodified
}
\examples{
\dontrun{
relu(c(0, -1, 3, -5, 7))
}
}
\author{
Yijun Yang
}
