% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relu_p.R
\name{relu_p}
\alias{relu_p}
\title{Derivative of ReLU}
\usage{
relu_p(v)
}
\arguments{
\item{v}{a numeric vector or matrix}
}
\value{
the original vector with positive values to 1 and negative values to zero
}
\description{
Apply derivative of the rectified linear unit (ReLU).
}
\examples{
\dontrun{
relu(c(0, -1, 3, -5, 7))
}
}
\author{
(refer to CASL p218)
}
