% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{ridge}
\alias{ridge}
\title{Ridge regression using SVD.
Source: Github called bis557, lecture notes, originally written by Michale Kane}
\usage{
ridge(form, data, lambda = 0)
}
\arguments{
\item{form}{The formula to extract X from input dataset.}

\item{data}{Input dataset.}

\item{lambda}{Penalization parameter, default=0}
}
\value{
The estimated model as a list.
}
\description{
Ridge regression using SVD.
Source: Github called bis557, lecture notes, originally written by Michale Kane
}
\examples{
ridge(Sepal.Length ~ ., iris, lambda = 0.1)



}
