% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_hw2d.R
\name{ridge_hw2d}
\alias{ridge_hw2d}
\title{Ridge Regression: Finding Lambda}
\usage{
ridge_hw2d(form, d, lambda_vals, contrasts = NULL)
}
\arguments{
\item{form}{linear model formula}

\item{d}{data frame}

\item{lambda_vals}{the lambda values (regularization parameter)}

\item{contrasts}{optional list of constants for factor variables aka contrast}
}
\description{
This function finds the lambda which gives the lowest loss.
}
\examples{
data(iris)
ridge_hw2d(form = Sepal.Length ~ ., d = iris[,-5],
           lambda_vals = 10^seq(-3, 2, length = 200))

}
