% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_lambda_optimization.R
\name{ridge_lambda_optimization}
\alias{ridge_lambda_optimization}
\title{Optimizing Ridge Regression Parameter Lambda}
\usage{
ridge_lambda_optimization(form, data, folds = 10, lambdas = seq(0, 0.5,
  0.01))
}
\arguments{
\item{form}{Formula of y~x1+x2+...}

\item{data}{The data set from which one desires to build the linear model}

\item{folds}{Folds refers to the number of groups that a given data sample should be split into}

\item{lambdas}{Lambda parameter to use for ridge regression}
}
\value{
One value providing the optimal lambda parameter to use for ridge regression
}
\description{
This is a method to optimize selection of the parameter lambda in ridge regression. Note:
a cross validation approach will be adopted to select the best value for lambda, which involves splitting
the data into a training and test set. We want a model to be fit to the training set with a specific value
of lambda, then determine values of beta coeffiecients and compare to those derived from the test data set.
We will repeat this process for many different values of lambda and choose the model with the highest
accuracy on the test set.
}
