% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_py_hw4a.R
\name{ridge_py_hw4a}
\alias{ridge_py_hw4a}
\title{OLS with Ridge Regression - Python}
\usage{
ridge_py_hw4a(y, X, lambda_val)
}
\arguments{
\item{y}{target/response variable matrix}

\item{X}{independent variables matrix (contrasts should be taken care of here)}

\item{lambda_val}{the lambda value (regularization parameter)}
}
\value{
the best-fit coefficients with ridge regularization
}
\description{
This function fits linear models with the ridge regression
regularization (using the l2 norm). This method is used to reduce overfitting.
Language is in Python 3.
}
\examples{
data(iris)
y <- matrix(iris$Sepal.Length, ncol = 1)
X <- model.matrix( ~ . - Sepal.Length - Species, data = iris)
ridge_py_hw4a(y, X, lambda_val = 1)

}
