% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_R.R
\name{ridge_R}
\alias{ridge_R}
\title{Ridge Regression Considering Collinearity}
\usage{
ridge_R(X, Y, lambda = 0.01)
}
\arguments{
\item{X}{a X model matrix created by function "model_matrices".}

\item{Y}{a Y vector created by function "model_matrices".}

\item{lambda}{a penalty parameter specified by the user (default=0.01).}
}
\value{
a list including the coefficient estimates.
}
\description{
This is a ridge regression function taking into account collinear or nearly collinear regression variables.
}
\examples{
data(iris)
fit <- ridge_regression(X, Y)
fit$coefficients
}
