% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_reg.r
\name{ridge_reg}
\alias{ridge_reg}
\title{Fit a Ridge Regression model with given lambda}
\usage{
ridge_reg(form, lambda, d)
}
\arguments{
\item{form}{formula of the regression model, where intercept term is dropped}

\item{lambda}{tuning parameter of ridge regression, controlling panalty}

\item{d}{a data.frame, with each column scaled to have mean 0 and unit vairance}
}
\value{
An Ridge object
}
\description{
This function use SVD decomposition to solve ridge regression.
}
\examples{
data(ridge_train)
ridge_train_scale <- as.data.frame(scale(ridge_train))
fit <- ridge_reg(y ~. - 1, 0.01, ridge_train_scale)
summary(fit)
}
