% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridge_regression}
\alias{ridge_regression}
\title{Ridge Regression}
\usage{
ridge_regression(formula, data, lambda = 0, contrasts = NULL, tol = 1e-08)
}
\arguments{
\item{formula}{a formula that accounts for the model of interest}

\item{data}{a dataframe that contains the variables of interest}

\item{lambda}{ridge penalty with default value 0}

\item{contrasts}{an list of contrast}

\item{tol}{a tolerance to detect collinearity}
}
\value{
A list of beta estimates and the formula.
}
\description{
Implement a ridge regression function taking into account colinear (or nearly colinear) regression variables.
}
\examples{
data(iris)
ridge_regression(Sepal.Length ~ ., iris, 0.01)
}
