% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression1.R
\name{ridge_regression1}
\alias{ridge_regression1}
\title{Caluculate coefficients using ridge regression and finding the best lambda for ridge regression using cross validation.}
\usage{
ridge_regression1(form, data, lambda = 10)
}
\arguments{
\item{form}{dependent variable ~ all independent variables}

\item{data}{the dataframe that will be used to caluculate coefficients}

\item{lambda}{A user supplied lambda sequence. Typical usage is to have the program compute its own lambda sequence based on nlambda and lambda.min.ratio.}
}
\value{
ret: ridge regression coefficients
}
\description{
This is a generic function: we use ridge regression to calculate the linear coeffients between dependent variable and independent wariables and finding the best lambda for ridge regression using cross validation.
}
\examples{
ridge_regression1(Sepal.Length ~., iris, lambda=10)
}
