% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.r
\docType{data}
\name{ridge_test}
\alias{ridge_test}
\title{Validation data for the 2nd homework vignette}
\format{An object of class \code{data.frame} with 200 rows and 5 columns.}
\usage{
ridge_test
}
\description{
In homework 2 we build a ridge regression model.
For illustration, we would use this data set to assess the prediction error of 
the ridge regreession model built on training data.
}
\keyword{datasets}
