% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.r
\docType{data}
\name{ridge_train}
\alias{ridge_train}
\title{Training data for the 2nd homework vignette}
\format{An object of class \code{data.frame} with 200 rows and 5 columns.}
\usage{
ridge_train
}
\description{
In homework 2 we build a ridge regression model. 
For illustration, we would build ridge regression using this data set 
and validate its prediction performance on another data set, ridge_test
}
\keyword{datasets}
