% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_mnist_test_doc.R
\docType{data}
\name{sign_mnist_test}
\alias{sign_mnist_test}
\title{sign language MNIST}
\source{
\url{https://www.kaggle.com/datamunge/sign-language-mnist}
}
\description{
MNIST testing dataset
}
\keyword{data}
