% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_mnist_train_doc.R
\docType{data}
\name{sign_mnist_train}
\alias{sign_mnist_train}
\title{sign language MNIST}
\source{
\url{https://www.kaggle.com/datamunge/sign-language-mnist}
}
\description{
MNIST training dataset
}
\keyword{data}
