% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softmax.R
\name{softmax}
\alias{softmax}
\title{Softmax Function}
\usage{
softmax(v)
}
\arguments{
\item{v}{a numeric vector}
}
\value{
a vector after applying the softmax function
}
\description{
apply a function to the output layer to make it behave as a proper set of probabilities
}
\examples{
\dontrun{
softmax(c(0, -1, 3, -5, 7))
}
}
\author{
Yijun Yang
}
