% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softmaxreg.R
\name{softmaxreg}
\alias{softmaxreg}
\title{Softmax regression with gradient descent}
\usage{
softmaxreg(X, y, lrate = 0.01, maxiter = 10000, tol = 1e-05)
}
\arguments{
\item{X}{the design matrix}

\item{y}{the response variable}

\item{lrate}{the learning rate}

\item{maxiter}{the maximum iteration number}

\item{tol}{the numerical tolerance}
}
\description{
Fits softmax regression using gradient descent, by maximizing log-likelihood
}
