% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-matrix.R
\docType{class}
\name{sparse.matrix-class}
\alias{sparse.matrix-class}
\title{Create a new class called "sparse.matrix", and define addition, multiplication, and transpose of it.}
\arguments{
\item{i}{a vector indicating row number with non-zero elements}

\item{j}{a vector indicating columnn number with non-zero elements}

\item{x}{a vector indicating the value of each non-zero element}

\item{dims}{a vector indicating dimensions of the matrix}

\item{e1}{a sparse.matrix}

\item{e2}{a sparse.matrix}

\item{x}{a sparse.matrix}

\item{y}{a sparse.matrix}
}
\value{
A sparse.matrix object
}
\description{
I set a new class called "sparse.matrix", and created a function also called 
sparse.matrix to return a "sparse.matrix" object. Then, I created functions sparse_add, 
sparse_multiply and sparse_tranpose to do sparse matrix operations. At last, I set methods 
"+", "%*%" and "t" by using the three functions.
}
\examples{
sm <- sparse.matrix(i = c(1, 2), j = c(1, 1), x = c(1, 1), dims = c(2, 2))
sm
sm + sm
sm \%*\% sm
t(sm)
}
