% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-matrix.R
\docType{methods}
\name{sparse.matrix}
\alias{sparse.matrix}
\alias{t,sparse.matrix-method}
\alias{t,sparse.matrix,ANY-method}
\alias{+,sparse.matrix,sparse.matrix-method}
\alias{\%*\%,sparse.matrix,sparse.matrix-method}
\title{An S4 class for sparse matrix.}
\usage{
sparse.matrix(i, j, x, dims = NULL)

\S4method{t}{sparse.matrix}(x)

\S4method{+}{sparse.matrix,sparse.matrix}(e1,e2)

\S4method{\%*\%}{sparse.matrix,sparse.matrix}(x, y)
}
\arguments{
\item{x, y, e1, e2, i, j, dims}{a \code{sparse.matrix} object}
}
\description{
An S4 class for sparse matrix.

t method for sparse.matrix

+ method for sparse.matrix

%*% method for sparse.matrix
}
\section{Slots}{

\describe{
\item{\code{i}}{row index}

\item{\code{j}}{colum index}

\item{\code{x}}{value}

\item{\code{dims}}{dimensions}
}}

