% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW4.R
\name{sparse.matrix}
\alias{sparse.matrix}
\title{create a new class of sparse matrix}
\usage{
sparse.matrix(i, j, x, dims = c(max(i), max(j)))
}
\arguments{
\item{i}{row index of non-zero element}

\item{j}{col index of non-zero element}

\item{x}{value of non-zero element corresponding to i, j}

\item{dims}{dimensions of the sparse matrix}
}
\value{
A sparse.matrix object
}
\description{
create sparse matrix using sparse.matrix, operations defined including add, multiply and transpose
}
