% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_computation.r
\name{sparse_add}
\alias{sparse_add}
\title{Add two sparse matrices}
\usage{
sparse_add(a, b)
}
\arguments{
\item{a}{a sparse matrix}

\item{b}{a sparse matrix}
}
\value{
the sum of a and b in the sparse.matrix form.
}
\description{
Add two sparse.matrix objects
}
