% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_computation.r
\name{sparse_multiply}
\alias{sparse_multiply}
\title{Multiply two sparse matrices}
\usage{
sparse_multiply(a, b)
}
\arguments{
\item{a}{a sparse matrix}

\item{b}{a sparse matrix}
}
\value{
the product of a and b in the sparse.matrix form
}
\description{
Multiply two sparse.matrix objects
}
