% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_computation.r
\name{sparse_transpose}
\alias{sparse_transpose}
\title{Transpose a sparse matricx}
\usage{
sparse_transpose(a)
}
\arguments{
\item{a}{a sparse.matrix object}
}
\value{
the trainsposed form of the sparse matrix
}
\description{
transpose a sparse matrix
}
