% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-matrix.R
\docType{methods}
\name{t,sparse.matrix-method}
\alias{t,sparse.matrix-method}
\title{Define transpose of sparse.matrix object.}
\usage{
\S4method{t}{sparse.matrix}(x)
}
\arguments{
\item{x}{a sparse.matrix object}
}
\description{
A method for sparse.matrix transpose.
}
