% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework-4.R
\name{t.sparse.matrix}
\alias{t.sparse.matrix}
\title{Define transpose method for 'sparse.matrix' class}
\usage{
\method{t}{sparse.matrix}(x)
}
\arguments{
\item{x}{A 'sparse.matrix' object.}
}
\value{
A 'sparse.matrix' object, which is the result of t(x).
}
\description{
This function is used for defining `t` operation in class "sparse.matrix".
}
