% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_computation.r
\name{to_sparse_matrix}
\alias{to_sparse_matrix}
\title{Convert a data.frame object into a sparse.matrix object}
\usage{
to_sparse_matrix(df, dims)
}
\arguments{
\item{df}{a data frame in the specific form: with (i,j,x) columns}

\item{dims}{the dimension of the underlying dense matrix}
}
\value{
a sparse.matrix object
}
\description{
Convert a data frame into a sparse.matrix object (because merge() is used in matrix algebra functions, 
which coerce matrices into dataframes)
}
