% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_matrix.r
\name{transpose}
\alias{transpose}
\title{Method of transpose}
\usage{
transpose(x, ...)
}
\arguments{
\item{m:}{a sparse matrix}
}
\value{
: transpose of a matrix
}
\description{
: transpose a sparse matrix
}
