% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_networks.R
\name{xy2pos}
\alias{xy2pos}
\title{A 3-layer Neural Networks}
\usage{
xy2pos(x, y, nrow)
}
\arguments{
\item{y}{input of y}

\item{X}{input of X}

\item{maxit}{maximum number of iterations}

\item{gamma}{The step size}

\item{H1}{number of nodes in the 1st layer}

\item{H2}{number of nodes in the 2nd layer}

\item{lambda}{penalty term}
}
\value{

}
\description{
Use a 3-layer artificial neural networks to do classification on sprial data
}
\examples{
\dontrun{
NN_classify(X, y, maxit = 1e5, gamma = 1e-2, H1 = 10, H2 = 10, lambda = 0.1)
}
}
\author{
Yijun Yang
}
