% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_lambda.R
\name{best_lambda}
\alias{best_lambda}
\title{Optimize the ridge parameter lambda with k-fold cross-validation}
\usage{
best_lambda(formula, data, nfold = 5, lambda_list, intercept = TRUE)
}
\arguments{
\item{formula}{An object of class formula to describe how the model is fitted}

\item{data}{A data frame}

\item{nfold}{An optional number k for k-fold cross-validation, default = 5}

\item{lambda_list}{A numeric vector consisting of lambda candidates}

\item{intercept}{An optional logical value to indicate if intercept is
included, default = TRUE}
}
\value{
Return the best lambda value
}
\description{
Optimize the ridge parameter lambda with k-fold cross-validation
}
\examples{
data(iris)
best_lambda(formula = Sepal.Length ~ .,
            data = iris, nfold = 5,
            lambda_list = 10^seq(-2, 2, by = .1))

data(mtcars)
best_lambda(formula = mpg ~ drat + wt + qsec,
            data = mtcars, nfold = 3,
            lambda_list = 10^seq(-2, 2, by = .1))
}
