% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q3_opmitize_lamda.R
\name{best_lambda}
\alias{best_lambda}
\title{optimizing the ridge parameter λ.}
\usage{
best_lambda(formula, data, lambdas, nsplit)
}
\arguments{
\item{formula}{The equation of the ridge regression model}

\item{data}{The data needed to be fitted by ridge model}

\item{lambdas}{a set of candidate lambda for optimizing}

\item{nsplit}{the fold for sampling}
}
\value{
the best lambda
}
\description{
This function is aimed to optimize the ridge parameter λ.
}
