% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn-backward-prop.R
\name{casl_nn_backward_prop}
\alias{casl_nn_backward_prop}
\title{Apply backward propagation algorithm.}
\usage{
casl_nn_backward_prop(x, y, weights, f_obj, sigma_p, f_p)
}
\arguments{
\item{x:}{A numeric vector representing one row of the input.}

\item{y:}{A numeric vector representing one row of the response.}

\item{weights:}{A list created by casl_nn_make_weights.}

\item{f_obj:}{Output of the function casl_nn_forward_prop.}

\item{sigma_p:}{Derivative of the activation function.}

\item{f_p:}{Derivative of the loss function.}
}
\value{
A list containing the derivative of the loss function
with respect to z (grad_z) and w (grad_w).
}
\description{
Backward propagation.
}
