% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casl-neural-net.R
\name{casl_nn_forward_prop}
\alias{casl_nn_forward_prop}
\title{Apply the Forward Propagation to a Set of Neural Network Weights and Biases (from CASL)}
\usage{
casl_nn_forward_prop(x, weights, sigma)
}
\arguments{
\item{x}{A numeric vector representing one row of the input}

\item{weights}{A list created by casl_nn_make_weights}

\item{sigma}{The activation function (such as ReLU)}
}
\value{
A list containing the new weighted responses (z) and activations (a)
}
\description{
This function (from CASL) implements the forward propagation of the inputs through a dense neural network given the network's weights and biases.
}
\examples{
# From page 215, in the code for the casl_nn_sgd function
x <- rnorm(1)
sizes <- c(1, 25, 1)
weights <- casl_nn_make_weights(sizes)
# We then do something to update the weights, like stochastic gradient descent (SGD)
f_obj <- casl_nn_forward_prop(x, weights, casl_util_ReLU)
}
