% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casl-neural-net.R
\name{casl_nn_make_weights}
\alias{casl_nn_make_weights}
\title{Create a Default List of Weights to Describe a Dense Neural Network (from CASL)}
\usage{
casl_nn_make_weights(sizes)
}
\arguments{
\item{sizes}{A vector of integers giving the size of each layer, including the input and output layers}
}
\value{
A list containing initialized weights and biases
}
\description{
This function (from CASL) creates a default list of weights for a dense neural network using a standard normal distribution.
}
\examples{
weights <- casl_nn_make_weights(sizes=c(1, 25, 1)) #From page 216
}
