% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn-predict.R
\name{casl_nn_predict}
\alias{casl_nn_predict}
\title{Predict values from a training neural network.}
\usage{
casl_nn_predict(weights, X_test)
}
\arguments{
\item{weights}{List of weights describing the neural network.}

\item{X_test}{A numeric data matrix for the predictions.}
}
\value{
A matrix of predicted values.
}
\description{
predict outcomes by neural network.
}
