% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casl-neural-net.R
\name{casl_nn_predict}
\alias{casl_nn_predict}
\title{Predict Values from a Training Neural Network (from CASL)}
\usage{
casl_nn_predict(weights, X_test)
}
\arguments{
\item{weights}{List of weights describing the neural network}

\item{X_test}{A numeric data matrix for the predictions}
}
\value{
A matrix of predicted values
}
\description{
This function (from CASL) uses the weights and biases calculated in a dense neural with using stochastic gradient descent (SGD) to calculate predicted values given new input data.
}
\examples{
# From pages 216-217
X <- matrix(runif(1000, min=-1, max=1), ncol=1)
y <- X[,1,drop = FALSE]^2 + rnorm(1000, sd = 0.1)
weights <- casl_nn_sgd(X, y, sizes=c(1, 25, 1), epochs=25, eta=0.01)
y_pred <- casl_nn_predict(weights, X)
}
