% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-mae-p.R
\name{casl_util_mae_p}
\alias{casl_util_mae_p}
\title{Derivative of the mean absolute error (MAE) function.}
\usage{
casl_util_mae_p(y, a)
}
\arguments{
\item{y}{A numeric vector of responses.}

\item{a}{A numeric vector of predicted responses.}
}
\value{
A numeric vector of the derivative.
}
\description{
Current derivative of the MAE function
}
