% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-mse-p.R
\name{casl_util_mse_p}
\alias{casl_util_mse_p}
\title{Derivative of the mean squared error (MSE) function.}
\usage{
casl_util_mse_p(y, a)
}
\arguments{
\item{y}{A numeric vector of responses.}

\item{a}{A numeric vector of predicted responses.}
}
\value{
A numeric vector of the derivative.
}
\description{
Current derivative of the MSE function
}
