% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casl-neural-net.R
\name{casl_util_mse_p}
\alias{casl_util_mse_p}
\title{Derivative of the Mean Squared Error (MSE) Function (from CASL)}
\usage{
casl_util_mse_p(y, a)
}
\arguments{
\item{y}{A numeric vector of responses}

\item{a}{A numeric vector of predicted responses}
}
\value{
A numeric vector of the current derivative of the MSE function
}
\description{
This function (from CASL) is the derivative of the mean squared error (MSE) loss function used for dense neural networks. We do not need the 1/n factor in front, because we will only end up using this function on scalar values.
}
\examples{
casl_util_mse_p(0, rnorm(1))
}
