% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ReLU-p.R
\name{casl_util_ReLU_p}
\alias{casl_util_ReLU_p}
\title{Apply derivative of the rectified linear unit (ReLU).}
\usage{
casl_util_ReLU_p(v)
}
\arguments{
\item{v}{A numeric vector or matrix}
}
\value{
A numeric vector or matrix
}
\description{
Set positive values to 1 and negative values to zero
}
