% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_ridge.R
\name{cv_ridge}
\alias{cv_ridge}
\title{Cross Validation for Ridge Regression}
\usage{
cv_ridge(form, d, folds = 10, lambdas)
}
\arguments{
\item{form}{a formula}

\item{d}{a dataframe}

\item{folds}{number of folds}

\item{lambdas}{a vector of ridge penalty term}
}
\value{
the best lambda for ridge regression and a plot
}
\description{
Optimize the ridge parameter.
}
\examples{
\dontrun{
cv_ridge(body_mass_g ~ bill_depth_mm, penguins, folds = 5, lambdas = seq(1, 10, 1))
}
}
\author{
Yijun Yang
}
