% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{cv_ridge_regression}
\alias{cv_ridge_regression}
\title{CV for Ridge Regression}
\usage{
cv_ridge_regression(formula, data, folds = 5, lambdas = seq(0, 1,
  0.03), contrasts = NULL)
}
\arguments{
\item{formula}{a model formula of the form "y ~ ..."}

\item{data}{a data frame}

\item{folds}{the number of folds on which to cross validate}

\item{lambdas}{a vector of lambdas to check}

\item{contrasts}{an optional list of contrasts}
}
\value{
A tibble containing summary statistics of RMSEs by lambda.
}
\description{
a minimal implementation to find optimal lambdas from a given set to use in ridge regression.
}
\details{
The bulk of the code is adapted from Professor Kane's example in class, which I then generalized to work within a function with some additional modifications.
}
