% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epan-kernel.R
\name{epan_kernel}
\alias{epan_kernel}
\title{Calculate the Epanechnikov Kernel}
\usage{
epan_kernel(x, h = 1)
}
\arguments{
\item{x}{A vector of imputs for this function}

\item{h}{The bandwidth, which has the default value of 1}
}
\value{
A numerical vector of K_h(x) evaluated at each value of x
}
\description{
This function returns a vector of values for the Epanechnikov Kernel that is the same length as the input vector. This implementation is based on the one in A Computational Approach to Statistical Learning, Pages 80-81.
}
\examples{
x <- -5:5
h <- 3
epan_kernel(x, 3)
}
