% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_order_glm.R
\name{first_order_glm}
\alias{first_order_glm}
\title{First-Order GLM Maximum Likelihood Solution}
\usage{
first_order_glm(
  X,
  y,
  mu_fun,
  maxit = 100,
  tol = 1e-10,
  size,
  adapt = FALSE,
  gamma = 0.9
)
}
\arguments{
\item{X}{A matrix of independent variables}

\item{y}{A vector of response variable}

\item{mu_fun}{The function specifying the mean function in exponential family}

\item{maxit}{Maximum number of iterations}

\item{tol}{Tolerance for the convergence}

\item{size}{A number specifying the constant step size in gradient descent}

\item{adapt}{A logical value. If TRUE, we use momentum gradient descent. If FALSE, we use constant step size}

\item{gamma}{A parameter controlling the momentum update}
}
\value{
A vector of fitted regression coefficients.
}
\description{
First-Order GLM Maximum Likelihood Solution
}
\examples{
\dontrun{first_order_glm(X, y,mu_fun = function(eta) exp(eta), size = 0.00001,adapt = FALSE,maxit = 1000)}
}
