% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_coeffs.R
\name{grab_coeffs}
\alias{grab_coeffs}
\title{Grab the slope coefficients
This is a simple function to grab the slope coefficients of a linear model}
\usage{
grab_coeffs(lm_ob)
}
\arguments{
\item{lm_ob}{the linear model object created by the lm() function}
}
\description{
Grab the slope coefficients
This is a simple function to grab the slope coefficients of a linear model
}
\examples{
library(palmerpenguins)
fit<-lm(penguins$bill_length_mm~., data=penguins[,-8])
grab_coeffs(fit)
}
