% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_coefficients.R
\name{grab_coeffs}
\alias{grab_coeffs}
\title{Grab the slope coefficients}
\usage{
grab_coeffs(lm_ob)
}
\arguments{
\item{lm_ob}{the linear object created by the lm() function.}
}
\description{
This is a simple function to grab the slope coefficients of
the linear model
}
\examples{
library(palmerpenguins)
fit<- lm(bill_length_mm~.,data=penguins[,-8])
grab_coeffs(fit)

}
