% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_resid.R
\name{grab_resids}
\alias{grab_resids}
\title{Grab the residuals
This is a simple function to grab the residuals of a linear model}
\usage{
grab_resids(lm_ob)
}
\arguments{
\item{lm_obj}{the linear model object created by the lm() function}
}
\description{
Grab the residuals
This is a simple function to grab the residuals of a linear model
}
\examples{
library(palmerpenguins)
fit<-lm(penguins$bill_length_mm~., data=penguins[,-8])
grab_resids(fit)
}
