% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad_descent.R
\name{grad_descent}
\alias{grad_descent}
\title{grad_descent() function}
\usage{
grad_descent(form, data, contrasts = NULL, itr = 1e+06)
}
\arguments{
\item{form}{a formula;}

\item{data}{a data frame used for the function;}

\item{contrasts}{a list of contrasts for factor variables}

\item{itr}{number of iteration}
}
\description{
another way to build a linear model. gradient descent for ordinary least squares
}
\examples{
data(iris)
fit_linear_model <- grad_descent(Sepal.Length ~ ., iris, contrasts = list(Species = "contr.sum"))
}
