% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient-descent.R
\name{gradient_descent}
\alias{gradient_descent}
\title{Implement Gradient Descent for OLS}
\usage{
gradient_descent(formula.input, data.input, contrasts.input = NULL,
  gamma.input = 1e-04, max.itera = 1e+06, stop.diff = 1e-12)
}
\arguments{
\item{formula.input}{a formula with the legal format.}

\item{data.input}{a dataframe provided by the user.}

\item{contrasts.input}{a list of contrasts of interest (default=NULL).}

\item{gamma.input}{a factor defined by user as a pace or speed of gradient descent (default=0.0001).}

\item{max.itera}{a maximum number of iterations set by user (default=1e6).}

\item{stop.diff}{a minimum absolute difference between SSR and the updated,
which is an alternative criterion to stop the iterations (default=1e-12).}
}
\value{
A list including the OLS estimates, simulating what will be produced by 'lm()'.
}
\description{
This is a function to numerically compute OLS estimates via
implementing the gradient descent method. Its reception of inputs and the
coefficient grabbing are all similar to 'lm()' function.
}
\examples{
data(iris)
fit <- linear_model(Sepal.Length ~ ., iris)
fit$coefficients
}
