% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient-descent.R
\name{gradient_descent}
\alias{gradient_descent}
\title{Implement Gradient Descent for OLS}
\usage{
gradient_descent(formula.input, data.input, contrasts.input = NULL,
  gamma.input = 1e-04, beta0 = 1, stop.diff = 1e-12,
  max.itera = 1e+06)
}
\arguments{
\item{formula.input}{a formula with the legal format.}

\item{data.input}{a dataframe provided by the user.}

\item{contrasts.input}{a list of contrasts of interest (default=NULL).}

\item{gamma.input}{a factor defined by user as a pace or speed of gradient descent (default=0.0001).}

\item{beta0}{an initializer for the beta vector that can be changed by user (default=1).}

\item{stop.diff}{an absolute difference between SSR and the updated one set by user
to stop the iteration at some satisfying point (default=1e-12).}

\item{max.itera}{a maximum number of iterations set by user,
which is an alternative criterion to stop the iterations (default=1e6).}
}
\value{
a list including the OLS estimates, simulating what will be produced by 'lm()'.
}
\description{
This is a function to numerically compute OLS estimates via
implementing the gradient descent method. Its reception of inputs and the
coefficients grabbed are all similar to 'lm()' function.
}
\examples{
data(iris)
fit <- linear_model(Sepal.Length ~ ., iris)
fit$coefficients
}
