% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_descent.R
\name{gradient_descent}
\alias{gradient_descent}
\title{fit a linear model}
\usage{
gradient_descent(formula, df, contrasts = NULL, gamma = 1e-04,
  iter = 10^5)
}
\arguments{
\item{formula}{an object of class "formula": describing the model to be fitted.}

\item{df}{a dataframe that containing all the variables in the model.}

\item{gamma}{gamma_k, the learning rate}

\item{iter}{number of iterations}

\item{contrast}{an optional list of contrasts for factor variables.}
}
\description{
This is a function taking a formula, a dataframe
and an optional list of contrast for factor variables,
and using gradient descent to return to coefficients.
with reference to https://towardsdatascience.com/linear-regression-using-gradient-descent-97a6c8700931
}
\examples{
library(palmerpenguins)
data(penguins)
form <- body_mass_g~.
gradient_descent(formula = form,data = penguins)
}
