% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_descent.R
\name{gradient_descent}
\alias{gradient_descent}
\title{Gradient descent}
\usage{
gradient_descent(form, dat, gamma = 0.1, tol = 1e-05, iter = 1e+05)
}
\arguments{
\item{form}{the input formula object}

\item{dat}{the input dataframe}

\item{gamma}{the learning rate}

\item{tol}{the numerical tolerance}

\item{iter}{the maximum iterations}
}
\description{
Fits linear model by gradient descent
}
