% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{gradient_descent}
\alias{gradient_descent}
\title{gradient descent}
\usage{
gradient_descent(formula, data, lr = 1e-04, max_Iter = 5e+05,
  threshold = 1e-12, print = FALSE)
}
\arguments{
\item{formula}{A formula of form "A ~ B".}

\item{data}{A dataframe.}

\item{lr}{learning rate by which to adjust beta along gradient;}

\item{max_Iter}{Maximum number of iterations;}

\item{threshold}{Size of difference between previous and new betas at which the loop exits; default is 0.001.}

\item{print}{Printing progress}
}
\description{
implementation of gradient descent for ordinary least squares, finding the solutions
for Y = X*beta.
}
\examples{
betas = gradient_descent(mpg ~ cyl + disp, mtcars)
}
