% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{gradient_descent}
\alias{gradient_descent}
\title{gradient descent}
\usage{
gradient_descent(X, Y, lambda = 0.01, N = 5000, print = FALSE)
}
\arguments{
\item{X}{the feature vector}

\item{Y}{the label vector}

\item{lambda}{learning rate}

\item{N}{number of iterations}

\item{print}{Used to hide or show the progress}
}
\description{
implementation of gradient descent for ordinary least squares, finding the solutions
for Y = X*beta.
}
\examples{
X = matrix(runif(9),nrow = 3, ncol = 3)
Y = as.matrix(c(1,2,3))
beta = gradient_descent(X,Y)
}
