% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_descent.R
\name{gradient_descent}
\alias{gradient_descent}
\title{Implement gradient descent for ordinary least squares}
\usage{
gradient_descent(form, data, alpha, num_iters, contrasts = NULL)
}
\arguments{
\item{form}{A given formula for fitting the regression}

\item{data}{A given dataset to fit the model}

\item{alpha}{The learning rate}

\item{num_iters}{The number of iterations}

\item{contrasts}{list of constasts for factor variables}
}
\description{
This is a function implementing gradient descent algorithm to
estimate the parameters in OLS model
}
\examples{
library(palmerpenguins)
data(penguins)
my_fit =gradient_descent(bill_length_mm ~ ., data = penguins[,-8],alpha=0.1,num_iters=1000)
}
