% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_descent.R
\name{gradient_descent}
\alias{gradient_descent}
\title{Gradient descent function for OLS regression}
\usage{
gradient_descent(
  formula,
  data,
  contrasts = NULL,
  alpha = 1e-04,
  tolerance = 1e-20
)
}
\arguments{
\item{formula}{Formula of the regression model}

\item{data}{The data used for gradient descent}

\item{contrasts}{A list of contrasts for factor variables.}

\item{alpha}{Learning rate}

\item{tolerance}{The threshold of iterations}
}
\value{
The estimated coefficients for OLS regression.
}
\description{
Gradient descent is a optimization algorithm which is often used to find the minimum of a function
}
\examples{
data(iris)
gradient_descent(Sepal.Length ~., iris, alpha=0.01, contrasts = NULL, tolerance=1e-10)
}
