% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_descent.R
\name{gradient_descent}
\alias{gradient_descent}
\title{Gradient Descent Function}
\usage{
gradient_descent(
  form,
  d,
  contrasts = NULL,
  iters = 1e+05,
  gamma = 1e-04,
  threshold = 1e-12
)
}
\arguments{
\item{form}{A formula}

\item{d}{A dataframe}

\item{contrasts}{Contrasts argument}

\item{iters}{Number of maximum iterations}

\item{gamma}{The learning rate}

\item{threshold}{Threshold for the difference between RSS of new and old betas.}
}
\value{

}
\description{
Implement gradient descent for ordinary least squares.
}
\examples{
\dontrun{
gradient_descent(body_mass_g ~ bill_depth_mm, penguins)
}
}
\author{
Yijun Yang
}
