% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_descent_new.R
\name{gradient_descent_new}
\alias{gradient_descent_new}
\title{Gradient descent based on the out-of-sample accuracy.}
\usage{
gradient_descent_new(
  formula,
  data_frame,
  contrasts = NULL,
  lambda = 1e-04,
  epsilon = 1e-20,
  iters = 1e+06
)
}
\arguments{
\item{formula}{The formula of the model to be fitted.}

\item{data_frame}{A data frame which contains the data for the model.}

\item{contrasts}{Default is NULL. A list of contrasts for factor variables.}

\item{lambda}{Default is 0.0001. The learning rate.}

\item{epsilon}{Default is 1e-20. The minimum difference between the current SSR and the updated SSR.}

\item{iters}{Default is 1e6. The maximum number of iterations.}
}
\value{
A list of estimated coefficients.
}
\description{
This is a function to fit the OLS model using gradient descent that calculates the penalty based on the out-of-sample accuracy.
}
\examples{
data(iris)
model <- gradient_descent_new(Sepal.Length ~ ., data = iris)
model$coefficients
}
