% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.matrix.R
\name{\%*\%.default}
\alias{\%*\%.default}
\title{Use for matrix multiple}
\usage{
\method{\%*\%}{default}(x, y)
}
\arguments{
\item{x}{A certain class matrix or vector object}

\item{y}{A certain class matrix or vector object}
}
\value{
A certain class matrix or vector object
}
\description{
This function is used for matrix multiple
}
