% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.matrix.R
\name{\%*\%.default}
\alias{\%*\%.default}
\title{Use for multipling two "sparse.matrix" objects}
\usage{
\method{\%*\%}{default}()
}
\arguments{
\item{a}{A "sparse.matrix object"}

\item{b}{A "sparse.matrix object"}
}
\value{
A "sparse.matrix" object
}
\description{
This function is used for multipling two "sparse.matrix" objects
}
\examples{
sm0 <- sparse.matrix(i = c(1, 2), j = c(1, 1), x = c(1, 1))
sm1 <- sparse.matrix(i = c(1, 2), j = c(1, 1), x = c(3, 1), dims = c(3, 2))
sm0 \%*\% sm1
}
