% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_matrix.r
\name{\%*\%.default}
\alias{\%*\%.default}
\title{Method of addition}
\usage{
\method{\%*\%}{default}()
}
\arguments{
\item{a:}{a sparse matrix}

\item{b:}{a sparse matrix}
}
\value{
: sum of two sparse matrices
}
\description{
: adds two sparse matrices
}
