% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.matrix_class.R
\name{\%*\%.sparse.matrix}
\alias{\%*\%.sparse.matrix}
\title{Matrix-Multiply Two sparse.matrix Objects}
\usage{
"\%*\%.sparse.matrix"(a, b)
}
\arguments{
\item{a}{A sparse.matrix object (constructed from a data.frame)}

\item{b}{A sparse.matrix object (constructed from a data.frame)}
}
\value{
A sparse.matrix object that is the matrix multiple of a and b
}
\description{
This method is the implementation of the sparse_multiply from the homework. It matrix-multiplies two sparse.matrix objects.
}
\examples{

}
