% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-matrix.R
\name{\%*\%.sparse.matrix}
\alias{\%*\%.sparse.matrix}
\title{Multiply-textbook p294-295}
\usage{
\method{\%*\%}{sparse.matrix}(a, b)
}
\arguments{
\item{a}{A list describing a sparse matrix.}

\item{b}{A list describing a sparse matrix.}
}
\description{
Function is used to multiply two sparse matrices stored as triplet lists.
}
\examples{
a <- sparse.matrix(i = c(1, 2), j = c(1, 1), x = c(3, 1))
b <- sparse.matrix(i = c(1, 2, 2), j = c(1, 1, 2), x = c(4.4, 1.2, 3))
sparse_multiply(a,b)
}
