% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.matrix.R
\name{\%*\%.sparse.matrix}
\alias{\%*\%.sparse.matrix}
\title{Use for multipling two "sparse.matrix" objects}
\usage{
\method{\%*\%}{sparse.matrix}(x, y)
}
\arguments{
\item{x}{A "sparse.matrix object"}

\item{y}{A "sparse.matrix object"}
}
\value{
A "sparse.matrix" object
}
\description{
This function is used for multipling two "sparse.matrix" objects
}
