% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_matrix.r
\name{\%*\%.sparse.matrix}
\alias{\%*\%.sparse.matrix}
\title{Method of multiplication}
\usage{
\method{\%*\%}{sparse.matrix}(m, n)
}
\arguments{
\item{m:}{a sparse matrix}

\item{n:}{a sparse matrix}
}
\value{
: product of two sparse matrices
}
\description{
: multiply two sparse matrices
}
