% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-matrix.R
\name{\%*\%.sparse.matrix}
\alias{\%*\%.sparse.matrix}
\title{Multiply two sparse matrices}
\usage{
"\%*\%.sparse.matrix"(a, b)
}
\arguments{
\item{a}{a sparse matrix}

\item{b}{a sparse matrix}
}
\value{
A sparse matrix
}
\description{
This function multiplies two sparse matrices.
}
\examples{
sm1 <- sparse.matrix(i = c(1, 2), j = c(1, 1), x = c(3, 1), dims = c(3, 2))
sm3 <- sparse.matrix(i = rep(1, 3), j = 1:3, x = 1:3, dims = c(2, 3))
sm1 \%*\% sm3
}
