% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.matrix.R
\name{\%*\%}
\alias{\%*\%}
\title{Use for matrix multiple based on the class of matrix}
\usage{
x \%*\% y
}
\arguments{
\item{x}{A certain class matrix or vector object}

\item{y}{A certain class matrix or vector object}
}
\value{
A certain class matrix or vector object
}
\description{
This function is used for matrix multiple
}
