% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epan-kernel.R
\name{kern_density}
\alias{kern_density}
\title{Calculate the Density Estimate using the Epanechnikov Kernel}
\usage{
kern_density(x, h, x_new)
}
\arguments{
\item{x}{A vector of points to use for training the density function}

\item{h}{The bandwidth for the underlying Epanechnikov Kernel, which has the default value of 1}

\item{x_new}{The vector of values on which to evaluate the density function}
}
\value{
A numerical vector of the density estimates of x_new using the Epanechnikov Kernel
}
\description{
This function returns a vector of values for the density function defined by the Epanechnikov Kernel.
}
\examples{
x <- -4:4
h <- 2
x_new <- -10:10
kern_density(x, h, x_new)
}
