% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-model.R
\name{linear_model}
\alias{linear_model}
\title{Fit a linear model}
\usage{
linear_model(formula, data, contrasts = NULL)
}
\arguments{
\item{formula}{a formula with the legal format.}

\item{data}{a dataframe provided by the user.}

\item{contrasts}{a list of contrasts of interest (default=NULL).}
}
\description{
This is a function implement the linear regression
and approximately reproduce the coefficients in the linear model
object that can be created by 'lm()'.
}
\examples{
data(iris)
linear_model(Sepal.Length ~ ., iris)
}
