% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-model.R
\name{linear_model}
\alias{linear_model}
\title{Fit a linear model}
\usage{
linear_model(formula.input, data.input, contrasts.input = NULL)
}
\arguments{
\item{formula.input}{a formula with the legal format.}

\item{data.input}{a dataframe provided by the user.}

\item{contrasts.input}{a list of contrasts of interest (default=NULL).}
}
\value{
a list including the OLS estimates, simulating what will be produced by 'lm()'.
}
\description{
This is a function to approximately reproduce the coefficients
in the linear model object that can be created by 'lm()'. The method being
used to compute the OLS estimates is orthogonal projection.
}
\examples{
data(iris)
fit <- linear_model(Sepal.Length ~ ., iris)
fit$coefficients
}
\references{
This work references the methodology and coding logic in the textbook
"A Computational Approach to Statistical Learning" by Arnold, Kane, and Lewis (p.24).
}
