% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Fitting Linear Model Function}
\usage{
linear_model(formula, data, contrasts = NULL)
}
\arguments{
\item{formula}{linear model formula}

\item{data}{data frame}

\item{contrasts}{optional list of constants for factor variables aka contrast}
}
\description{
This function fits linear models by calculating regressions.
}
\examples{
linear_model(Sepal.Length ~ ., iris)
linear_model(Sepal.Length ~ ., iris,
             contrasts = list(Species = "contr.sum"))

}
