% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Fitting Linear Model Function}
\usage{
linear_model(form, d, contrasts = NULL)
}
\arguments{
\item{form}{linear model formula}

\item{d}{data frame}

\item{contrasts}{optional list of constants for factor variables aka contrast}
}
\value{
the best-fit coefficients for linear regression
}
\description{
This function fits linear models by calculating regressions.
}
\examples{
data(iris)
linear_model(Sepal.Length ~ ., iris)
linear_model(Sepal.Length ~ ., iris,
             contrasts = list(Species = "contr.sum"))

}
