% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.r
\name{linear_model}
\alias{linear_model}
\title{fit a linear model}
\usage{
linear_model(formula, df, contrasts = NULL)
}
\arguments{
\item{formula}{an object of class "formula": describing the model to be fitted.}

\item{df}{a dataframe that containing all the variables in the model.}

\item{contrast}{an optional list of contrasts for factor variables.}
}
\description{
This is a function taking a formula, a dataframe
and an optional list of contrast for factor variables,
and fit a linear model and return the coefficients of the model.
}
\examples{
library(palmerpenguins)
data(penguins)
form <- body_mass_g~.
linear_model(formula = form,data = penguins)
}
