% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-model.r
\name{linear_model}
\alias{linear_model}
\title{fit the linear model}
\usage{
linear_model(formula, data_frame, contrasts = NULL)
}
\arguments{
\item{formula}{a formula of linear model}

\item{data_frame}{a data_frame}

\item{contrasts}{default is NULL. a list of constasts for factor variables}
}
\description{
This a simple function to fit the linear model.
}
\examples{
data(iris)
linear_model(Sepal.Length ~ ., iris,contrasts = NULL)$coefficients
}
