% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Fits linear model}
\usage{
linear_model(form, dat, contrasts = NULL)
}
\arguments{
\item{form}{the input formula object}

\item{dat}{the input dataframe}

\item{constrasts}{list of contrasts for factor input}
}
\description{
Fits linear model by solving normal equation
}
