% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Fits linear model}
\usage{
linear_model(f, df, constrasts = NULL)
}
\arguments{
\item{f}{the formula}

\item{df}{the dataframe}

\item{constrasts}{list of variables as input}
}
\description{
Fits linear model using built-in lm function
}
