% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-model.r
\name{linear_model}
\alias{linear_model}
\title{Preform Linear Model in Linear Algebra Way}
\usage{
linear_model(form, data, contrasts = NULL)
}
\arguments{
\item{form}{a formula object}

\item{data}{a data frame}

\item{contrasts}{a list, whose entries are values (numeric matrices, functions
or character strings naming functions) to be used as replacement values for
the contrasts replacement function and whose names are the names of columns
of data containing factors. If \code{NULL}, then contrasts are ommitted.}
}
\description{
This is a simple function to build a linear model.
}
\examples{
data(lm_patho)
fit <- linear_model(y ~ ., data = lm_patho)
fit$coefficients
}
