% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{linear_model}
\alias{linear_model}
\title{linear model function}
\usage{
linear_model(formula, data, contrasts = NULL)
}
\arguments{
\item{formula}{a formula for a linear model to regress.}

\item{data}{the data frame we run our model on.}

\item{contrasts}{the categorical variables we want to include in our model.}
}
\description{
This is a simple function to calculate the coefficients of of a linear model.
}
\examples{
library(palmerpenguins)
data(penguins)
coeffs = linear_model(bill_length_mm ~ ., data = penguins[,-8])
}
