% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Run a linear model}
\usage{
linear_model(form, dataframe, contrasts = NULL)
}
\arguments{
\item{dataframe}{data frame}

\item{contrasts}{optional; list of constrasts for factor variables}

\item{formula}{linear regression formula}
}
\description{
This function will run a linear model and return the coefficients
}
\examples{
data(iris)
linear_model(Sepal.Length ~ ., iris,contrasts = list(Species = "contr.sum"))
}
