% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Fit a linear model.}
\usage{
linear_model(formula, data_frame, contrasts = NULL)
}
\arguments{
\item{formula}{The formula of the regression model to be fitted.}

\item{data_frame}{A data frame which contains the data for the model.}

\item{contrasts}{A list of contrasts for factor variables.}
}
\value{
A list of estimated coefficients.
}
\description{
This a function to fit the linear model.
}
\examples{
data(iris)
model <- linear_model(Sepal.Length ~., iris, contrasts = list(Species = "contr.sum"))
model$coefficients
}
