% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-models.r
\name{linear_model}
\alias{linear_model}
\title{Fit a linear model}
\usage{
linear_model(formula, data, tol = 1e-16)
}
\arguments{
\item{formula}{a formula}

\item{data}{a data.frame}

\item{tol}{the tolerance of double precision floating point arithmetic, which is by default 1e-16 (according to the book)}
}
\value{
An lm object, but only the coefficients (not the standard errors, t-values, or two-sided probabilities), so don't use summary() on this!
}
\description{
This function is my own implementation of the "lm" function in R. The implementation is based on the SVD method from Chapter 2 of A Computational Approach to Statistical Learning.

This method only uses the singular values that define an inverse condtion number greater than the tolerance of double precision floating point arithmetic, which we assume is 1e-16.
}
\examples{
fit <- linear_model(Sepal.Length ~., iris)
fit
}
