% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-models.r
\name{linear_model}
\alias{linear_model}
\title{Fit a linear model}
\usage{
linear_model(formula, data)
}
\arguments{
\item{formula}{a formula}

\item{data}{a data.frame}
}
\value{
An lm object, but only the coefficients (not the standard errors, t-values, or two-sided probabilities), so don't use summary() on this!
}
\description{
This function is my own implementation of the "lm" function in R. The implementation is based on the SVD method from Chapter 2 of A Computational Approach to Statistical Learning.

This method only uses the covariates that are linearly independent, using the "alias" function from the "stats" package to perform this test.
}
\examples{
fit <- linear_model(Sepal.Length ~., iris)
fit
}
