% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-models.r
\name{linear_model}
\alias{linear_model}
\title{Fit a linear model}
\usage{
linear_model(formula, data)
}
\arguments{
\item{formula}{a formula}

\item{data}{a data.frame}
}
\value{
An lm object, but only the coefficients (not the standard errors, t-values, or two-sided probabilities)
}
\description{
This function is my own implementation of the "lm" function in R. It tests for and removes colinear variables from the model.
}
\examples{
fit <- linear_model(Sepal.Length ~., iris)
summary(fit)
}
