% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Linear Model}
\usage{
linear_model(form, data, contrasts = NULL)
}
\arguments{
\item{form}{A given formula for fitting the regression.}

\item{data}{A given dataset to fit the model.}

\item{contrasts}{a list of constasts for factor variables.}
}
\description{
This is a function takes a formula and a dataset as input,
fit a linear model based on the input and returns a list of coefficients
}
\examples{
library(palmerpenguins)
data(penguins)
my_fit = linear_model(bill_length_mm ~ ., data = penguins[,-8])
}
