% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{This function will fit a linear model.}
\usage{
linear_model(form, d, contrasts = NULL)
}
\arguments{
\item{form}{A formula specifying the model.}

\item{d}{A dataframe.}

\item{contrasts}{A list specifying the contrasts. The default is NULL.}
}
\value{
A list with fitted coefficients.
}
\description{
This function will fit a linear model.
}
\examples{
linear_model(Sepal.Length~.,iris)
}
