% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Linear Model}
\usage{
linear_model(formula, data, contrasts = NULL)
}
\arguments{
\item{formula}{A formula of form "A ~ B".}

\item{data}{A dataframe.}

\item{contrasts}{A list of contrasts.}
}
\value{
A list of coefficients from an lm model. The output is computed using the QR decomposition with the coefficients of collinear variables set to NA.
}
\description{
Compute a linear regression.
}
\examples{
\dontrun{
data(iris)
linear_model(Sepal.Length ~ ., iris, contrasts)
}
}
