% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Fit linear models by the linear_model() function}
\usage{
linear_model(formula, data, contrasts = NULL)
}
\arguments{
\item{formula}{The formula of the regression model}

\item{data}{A data frame that's being used in this model}

\item{contrasts}{A list of constasts for factor variables.}
}
\value{
A list containing formula, estimated coefficients and contrasts list.
}
\description{
This linear_model function aims to fit OLS linear regressions
}
\examples{
data(iris)
fit_linear_model <- linear_model(Sepal.Length ~., iris, contrasts = list(Species = "contr.sum"))
}
