% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{linear_model() function}
\usage{
linear_model(form, dataframe, contrasts = NULL)
}
\arguments{
\item{form}{a formula;}

\item{dataframe}{a data frame;}

\item{contrasts}{a list of contrasts for factor variables}
}
\description{
one way to build a linear model
}
\examples{
data(iris)
fit_linear_model <- linear_model(Sepal.Length ~ ., iris, contrasts = list(Species = "contr.sum"))
}
