% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Linear Model Function}
\usage{
linear_model(form, d, contrasts = NULL)
}
\arguments{
\item{form}{A formula}

\item{d}{A dataframe}

\item{contrasts}{Contrasts argument}
}
\value{
A list of coefficients
}
\description{
This is the 1st homework assignment of BIS557.
}
\examples{
\dontrun{
linear_model(body_mass_g ~ bill_depth_mm, penguins)
}
}
\author{
Yijun Yang
}
