% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{linear_model}
\usage{
linear_model(form, dat, contrasts = NULL)
}
\arguments{
\item{form}{A formula with the format of "Y ~ ."}

\item{dat}{A dataframe.}

\item{contrasts}{A list of contrasts.}
}
\value{
A list of coefficients from an linear regression model.
}
\description{
Compute a linear regression.
}
\examples{
linear_model(Sepal.Length ~ ., iris, contrasts)

}
