% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Linear Model}
\usage{
linear_model(formula, data, contrasts = NULL)
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted. This should be a formula class argument.}

\item{data}{Specification of a dataframe that contains the variables in the model.}

\item{contrasts}{A list of contrasts.}
}
\value{
A list of component that imitates the output of lm() function. Including estimated coefficients for predictors specified in the formula.
}
\description{
Implement a linear regression based on given model and data.
}
\examples{
data(iris)
linear_model(Sepal.Length ~ ., iris, contrasts)
}
