% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_patho-doc.r
\docType{data}
\name{lm_patho}
\alias{lm_patho}
\title{Artificial data for function testing}
\description{
An originally csv formatted dataset with 2 observations and 3 variables
for testing the functions in homework#1.
}
\keyword{data}
