% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lm_patho}
\alias{lm_patho}
\title{Linear Regression with 2 Predictors}
\format{
A data frame with 2 rows and 3 variables:
\describe{
  \item{y}{response variable}
  \item{x1}{independent variable 1}
  \item{x2}{independent variable 2}
}
}
\usage{
lm_patho
}
\description{
A linear regression dataset - 2 observations using 2 predictors
Note: This "system" of equations is very ill-conditioned (high condition #).
}
\keyword{datasets}
