% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_patho-doc.r
\docType{data}
\name{lm_patho}
\alias{lm_patho}
\title{The document for the lm_patho dataset}
\description{
The lm_patho data set of the first homework of BIS557.
}
\references{
See the lm_patho.csv for details.
}
\keyword{data}
