% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_patho.R
\docType{data}
\name{lm_patho}
\alias{lm_patho}
\title{A simple dataset used for linear regression}
\format{
A data frame with 2 rows and 3 variables:
\describe{
\item{y}{dependent variable}
\item{x1,x2}{independent variable}
}
}
\usage{
lm_patho
}
\description{
A dataset with three variables (y, x1, and x2) and two rows of observations
}
\keyword{datasets}
