% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{lm_patho}
\alias{lm_patho}
\title{The document for the lm_patho dataset}
\description{
The lm_patho dataset is an example dataset for the linear model.
}
\references{
lm_patho.csv
}
\keyword{data}
