% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lm_patho}
\alias{lm_patho}
\title{Sample dataset for testing regression functions.}
\format{An object of class \code{data.frame} with 2 rows and 3 columns.}
\usage{
lm_patho
}
\description{
Sample dataset for testing regression functions.
}
\keyword{datasets}
