% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_patho-doc.R
\docType{data}
\name{lm_patho}
\alias{lm_patho}
\title{Data from the IMAP study}
\description{
This dataset is a test set for linear_model function
}
\references{
Refer to the first homework description of BIS557
}
\keyword{data}
