% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lm_patho}
\alias{lm_patho}
\title{lm_patho}
\format{A data frame with 2 rows and 3 columns.
\describe{
 \item{y}{response variable}
 \item{x1}{explanatory variable}
 \item{x2}{explanatory variable}
  ...
}}
\usage{
lm_patho
}
\description{
A dataset that contrains y and x.
}
\keyword{datasets}
