% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lm_patho}
\alias{lm_patho}
\title{Testing the linear model}
\format{
A data frame with 2 rows and 2 variables
\describe{
\item{y}{response variable}
\item{x1,x2}{regressors}
...
}
}
\usage{
lm_patho
}
\description{
This dataset is used to test the linear_model() function.
}
\keyword{datasets}
