% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lm_patho}
\alias{lm_patho}
\title{Sample dataset for testing.}
\format{A data frame with 2 rows and 3 variables.
\describe{
  \item{y}{response}
  \item{x1}{predictor 1}
  \item{x2}{predictor 2}
}}
\usage{
lm_patho
}
\description{
A dataset for testing tougher cases with the linear_model function in this package`.
}
\keyword{datasets}
