% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_patho_documentation.R
\docType{data}
\name{lm_patho}
\alias{lm_patho}
\title{lm_patho dataset}
\format{
A data frame with 2 rows and 3 variables.
\describe{
  \item{y}{response}
  \item{x1}{predictor 1}
  \item{x2}{predictor 2}
}
}
\usage{
lm_patho
}
\description{
A CSV file which contains 2 independent variables (x1,x2) and 1 dependent variable (y) with 3 observations
}
\keyword{datasets}
