% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.r
\docType{data}
\name{lm_patho}
\alias{lm_patho}
\title{A data set for the first homework}
\description{
In homework 1 we explore the effect of numerical instability
on linear regressions. This data set gives us a basis for exploration.
}
