% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_multiclass.R
\name{logit_multiclass}
\alias{logit_multiclass}
\title{A logistic model works for more than 2 classes}
\usage{
logit_multiclass(X, y, iters = 25, epsilon = 1e-10)
}
\arguments{
\item{X}{data frame or matrix of all the independent variables (all the elements should be numetric)}

\item{y}{vector of the dependent variable}

\item{iters}{maximum number of iterations}

\item{epsilon}{error tolerence}
}
\description{
This function implement a logistic model which can make prediction of more than 2 classes
}
\examples{
X = rbind(matrix(rnorm(1000),200),matrix(rnorm(1000,5,3),200),matrix(rnorm(1000,10,1),200))
X = scale(X)
X = cbind(1,scale(X))
y = c(rep(1,200),rep(2,200),rep(3,200))
logit_multiclass(X,y)
}
