% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_model_matrices.R
\name{make_model_matrices}
\alias{make_model_matrices}
\title{Create model matrices}
\usage{
make_model_matrices(form, d)
}
\arguments{
\item{form}{a formula}

\item{d}{the data to be fit (a data frame)}
}
\description{
This function creates a model matrix using gradient descent. Written in class 09/30/20. NOTE - it removes all NAs
}
\examples{
data("iris")
irisform <- Sepal.Length ~.
make_model_matrices(irisform, iris)
}
