% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-logistic.R
\name{multiclass_logistic}
\alias{multiclass_logistic}
\title{Multi-Class Model Generalizing Logistic Regression}
\usage{
multiclass_logistic(form, data, maxit)
}
\arguments{
\item{form}{a formula with the legal format.}

\item{data}{a dataframe provided by the user.}

\item{maxit}{the number of iteration for the second-order gradient descent.}
}
\value{
a list including the dependent variable of interest, coefficient estimates, and membership propensities.
}
\description{
This function implement a classification model
generalizing logistic regression to accommodate more than two classes.
}
\examples{
q3.data <- penguinsi
form <- species ~ bill_length_mm + bill_depth_mm
q3_fit <- multiclass_logistic(form, q3.data, maxit = 50)
}
