% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_Ridge.R
\name{my_Ridge}
\alias{my_Ridge}
\title{Fits Ridge regression model}
\usage{
my_Ridge(X, y, lambda)
}
\arguments{
\item{X}{the data design matrix}

\item{y}{the response vector}

\item{lambda}{the penalty strength}
}
\description{
Implements Ridge regression model by SVD
}
