% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ridge.R
\name{my_ridge}
\alias{my_ridge}
\title{Ridge regression for collinear/near-collinear columns}
\usage{
my_ridge(f, d, lambda)
}
\arguments{
\item{f}{a formula the data will be fit too}

\item{lambda}{the adjustment/"penalty" parameter}
}
\description{
This function fits a ridge regression model, accounting for collinear/near collinear columns.
}
\examples{
data("iris")
irisform <- Sepal.Length ~.
my_ridge(irisform, iris, .02)
}
