% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_lambda.R
\name{opt_lambda}
\alias{opt_lambda}
\title{Implment cross validation for optimizing the ridge parameter}
\usage{
opt_lambda(X, y, n_folds = 5, lambda_vals = exp(seq(-2, 5, 0.1)))
}
\arguments{
\item{X}{numeric data matrix of predictors in a ridge regression model}

\item{y}{the response vector}

\item{n_folds}{the number of cross validation folds}

\item{lambda_vals}{a list of hyperparameter that controls the penalty term}
}
\value{
The optimal lambda and coefficients of the fitted ridge regression model
}
\description{
This function can select the optimal lambda in ridge regression and export coefficients
}
