% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_lambda.R
\name{optimize_lambda}
\alias{optimize_lambda}
\title{Optimize ridge parameter lambda}
\usage{
optimize_lambda(form, dataframe)
}
\arguments{
\item{form}{ridge regression formula}

\item{dataframe}{data frame}
}
\description{
This function optimizes the ridge parameter lambda and output
the coefficients and chosen lambda
}
\examples{
data(iris)
optimize_lambda(Sepal.Length ~ ., iris[,-5])
}
