% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_lambda.R
\name{optimize_lambda}
\alias{optimize_lambda}
\title{Optimize Lambda for Ridge Regression}
\usage{
optimize_lambda(formula, data, lambda_vals = seq(0, 1, by = 0.005),
  k = 10)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted.}

\item{data}{an data frame containing the variables in the model.}

\item{lambda_vals}{the search range of lambda values. The default value is a sequence from 0 to 1 by 0.005.}

\item{k}{the number of folds used in the k-fold cross validation. The default value is 10.}
}
\value{
The optimal lambda that gives minimum prediction error using cross validation.
}
\description{
This function finds the optimial hyperparameter lambda for ridge regression.
}
\examples{
data(iris)
scaled_iris <- as.data.frame(scale(iris[, -5], TRUE, TRUE))
opt_lambda <- optimize_lambda(Sepal.Length ~ ., scaled_iris,
      lambda_vals = seq(0, 1, by = 0.05), k = 10)

}
