% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_ridge.R
\name{optimize_ridge}
\alias{optimize_ridge}
\title{Optimizing the ridge parameter.}
\usage{
optimize_ridge(X, Y, lambdas, k)
}
\arguments{
\item{X}{The input design matrix.}

\item{Y}{The input response vector.}

\item{lambdas}{The list of input penalty parameters.}

\item{k}{An optional number k for k-fold cross-validation.}
}
\value{
The most optimal ridge parameter.
}
\description{
This is a function to optimize the ridge penalty parameter.
}
\examples{
  n <- 500
  p <- 5
  beta <- c(1,-1,0,0,2)
  set.seed(9999)
  X <- matrix(rnorm(n * p), ncol = p)
  alpha <- 0.05
  X[,1] <- X[,1] * alpha + X[,2] * (1 - alpha)
  Y <- X \%*\% beta + rnorm(n)
  lambdas <- seq(0.1,10, by=0.1)
result <- optimize_ridge(X, Y, lambda=seq(0.1,15,length.out=50),k=5)
print(result)
}
