% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penguinsi-doc.R
\docType{data}
\name{penguinsi}
\alias{penguinsi}
\title{penguinsi-doc}
\description{
The penguins data set after imputing all missing values using the package missForest
}
\references{
See the palmerpenguins::penguins documentation for information about the values
}
\keyword{data}
