% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penguinsi-doc.r
\docType{data}
\name{penguinsi}
\alias{penguinsi}
\title{Data from the IMAP study}
\description{
The penguins data set after imputing all miss values using the {missForest} package.
}
\references{
See the Palmerpenguins::penguins documentation for information about the values.
}
\keyword{data}
