% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-matrix.R
\name{+.sparse.matrix}
\alias{+.sparse.matrix}
\title{Add two sparse matrices}
\usage{
\method{+}{sparse.matrix}(a, b)
}
\arguments{
\item{a}{a sparse matrix}

\item{b}{a sparse matrix}
}
\value{
A sparse matrix
}
\description{
This function add two sparse matrices.
}
\examples{
sm0 <- sparse.matrix(i = c(1, 2), j = c(1, 1), x = c(1, 1))
sm0 + sm0
}
