% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeregression.R
\name{predict.ridge_regression}
\alias{predict.ridge_regression}
\title{predict.ridge_regression}
\usage{
\method{predict}{ridge_regression}(object, ...)
}
\arguments{
\item{object}{Object of class inheriting from 'ridge_regression'.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A vector of predictions or a matrix of predictions.
}
\description{
This is a function called predict.ridge_regression,
which predict values based on ridge regression model object.
}
