% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{predict.ridge_regression}
\alias{predict.ridge_regression}
\title{Prediction function for Ridge Regression function}
\usage{
\method{predict}{ridge_regression}(object, ...)
}
\arguments{
\item{object}{ridge_regression object}

\item{...}{first argument should be the data frame}
}
\description{
Prediction function for Ridge Regression function
}
