% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{predict.ridge_regression}
\alias{predict.ridge_regression}
\title{Prediction method for ridge regression}
\usage{
\method{predict}{ridge_regression}(object, ...)
}
\arguments{
\item{object}{ridge_regression object}

\item{...}{`(dataframe)`}
}
\value{
An estimation of coefficients based on test data.
}
\description{
Prediction method for ridge regression
}
