% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{predict.ridge_regression}
\alias{predict.ridge_regression}
\title{the predict method}
\usage{
\method{predict}{ridge_regression}(object, ...)
}
\arguments{
\item{object}{an object of ridge_regression type}

\item{...}{`(dataframe)`}
}
\value{
a prediction of Y for the given X
}
\description{
the predict method
}
