% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{predict.ridge_regression}
\alias{predict.ridge_regression}
\title{Predict Method for Ridge Regression}
\usage{
\method{predict}{ridge_regression}(object, ...)
}
\arguments{
\item{object}{ridge_regression object}

\item{...}{`(dataframe)`}
}
\value{
An estimate of new Y given X.
}
\description{
Predict Method for Ridge Regression
}
