% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{predict.ridge_regression}
\alias{predict.ridge_regression}
\title{Prediction for Ridge Regression}
\usage{
\method{predict}{ridge_regression}(object, ...)
}
\arguments{
\item{object}{ridge_regression object}

\item{...}{\code{(dataframe)}}
}
\value{
estimated Y given X
}
\description{
Return the predicition of Y given X via ridge regression.
}
\author{
Yijun Yang
}
