% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ridge.R
\name{predict_ridge}
\alias{predict_ridge}
\title{Prediction function for class my_ridge}
\usage{
predict_ridge(object, ...)
}
\arguments{
\item{object}{the my_ridge object}

\item{...}{the data frame used to fit the object}
}
\description{
This function predicts a value of Y based on my_ridge
}
\examples{
data("iris")
irisdf <- as.data.frame(iris)
irissamdf <- irisdf[sample(nrow(irisdf), 0.2*nrow(irisdf)), ]
irisform <- Sepal.Length ~.
obj <- my_ridge(irisform, iris, 0.0001, 100000)
predict_ridge(obj, irisdf)
}
