% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{ridge}
\alias{ridge}
\title{ridge regression function}
\usage{
ridge(formula, df, contrasts = NULL, lambda = 0)
}
\arguments{
\item{formula}{an object of class "formula": describing the model to be fitted.}

\item{df}{a dataframe that containing all the variables in the model.}

\item{contrasts}{an optional list of contrasts for factor variables.}

\item{lambda}{an optional penalty term for ridge regression, default is 0}
}
\description{
This is a function taking a formula, a dataframe
and an optional list of contrast for factor variables,
and taking into account colinear (or nearly colinear) regression variables and returning coefficients.
}
\examples{
data(iris)
form <- Sepal.Length ~ Sepal.Width
ridge(formula = form,df = iris)

}
