% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridge}
\alias{ridge}
\title{A ridge regression function}
\usage{
ridge(X, Y, lambda)
}
\arguments{
\item{X}{The predictors of the ridge regression model}

\item{Y}{The response variable of the ridge regression model}

\item{lambda}{The penalizing parameter}
}
\value{
The estimated coefficients for ridge regression and the penalizing parameter lambda
}
\description{
A ridge regression that takes into account colinear (or nearly colinear) regression variables
}
\examples{
X = matrix(c(rep(1, 10), seq(2,20,2)+rnorm(10)), ncol=2)
y = c(seq(10,100,10)+rnorm(10))
ridge(X, y, lambda=0.01)
}
