% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge-reg.R
\name{ridge_reg}
\alias{ridge_reg}
\title{Fit a ridge regression model}
\usage{
ridge_reg(formula, data, lambda)
}
\arguments{
\item{formula}{a formula}

\item{data}{a data.frame}

\item{lambda}{a lambda value for penalty}
}
\value{
A ridge_reg object
}
\description{
This function passes parameters to the ridge_reg function.
}
\examples{
ridge_fit <- ridge_reg(Sepal.Length ~., iris)
ridge_fit$coef
}
