% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge-reg.r
\name{ridge_reg}
\alias{ridge_reg}
\title{Fit a ridge regression model}
\usage{
ridge_reg(formula, lambda, data)
}
\arguments{
\item{formula}{a formula}

\item{lambda}{a ridge regression parameter}

\item{data}{a data.frame}
}
\value{
A ridge regression object
}
\description{
This function passes parameters to coefficients in ridge regression.
}
\examples{
fit_ridge <- ridge_reg(y ~. -1, lambda, ridge_train_scale)
summary(fit_ridge)
}
