% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge-regression.R
\name{ridge_reg}
\alias{ridge_reg}
\title{Fit a ridge regression model}
\usage{
ridge_reg(form, lambda, d)
}
\arguments{
\item{lambda}{a number}

\item{formula}{a formula}

\item{data}{a data.frame}
}
\value{
An list of rigde regression information
}
\description{
This function fits the ridge regression model.
}
\examples{
ridge_fit <- ridge_reg(Sepal.Length ~. , lambda = 1.2, iris)
}
