% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge-reg.R
\name{ridge_reg}
\alias{ridge_reg}
\title{Fit a ridge regression model}
\usage{
ridge_reg(formula, lambda, data)
}
\arguments{
\item{formula}{a formula}

\item{lambda}{a penalizing parameter lambda}

\item{data}{a data.frame}
}
\value{
an ridge_reg object
}
\description{
This function fits a ridge regression model.
}
\examples{
fit <- ridge_reg(Sepal.Length ~., iris, lambda = 1)
summary(fit)
}
