% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge-regression.r
\name{ridge_reg}
\alias{ridge_reg}
\title{Fit a Ridge Regression Linear Model}
\usage{
ridge_reg(formula, lambda, dat)
}
\arguments{
\item{formula}{a formula}

\item{lambda}{a non-negative lambda value for the ridge regression (0 just gives ordinary least squares regression)}

\item{dat}{a data.frame}
}
\value{
A ridge_reg object, which is a list containing a vector of named coefficients, the lambda value, and the formula
}
\description{
This function computes a ridge regression linear model. The implementation uses SVD.
}
\examples{
fit_reg <- ridge_reg(Sepal.Length ~., 1, iris)
fit_reg
}
