% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridge_reg}
\alias{ridge_reg}
\title{Fit a Ridge Regression Model}
\usage{
ridge_reg(formula, lambda, data)
}
\arguments{
\item{formula}{A formula}

\item{lambda}{A penalty value}

\item{data}{A data.frame}
}
\value{
A ridge_reg object
}
\description{
This function passes parameters to the ridge_reg function.
}
\examples{
fit <- ridge_reg(Sepal.Length ~., lambda = 1, iris)
summary(fit)
}
