% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_model.R
\name{ridge_reg}
\alias{ridge_reg}
\title{Fit a ridge regression model}
\usage{
ridge_reg(formula, data, lambda)
}
\arguments{
\item{formula}{a formula}

\item{data}{a data.frame}
}
\value{
An ridge regression object
}
\description{
This function passes parameters to the ridge regression function.
}
