% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_reg.R
\name{ridge_reg}
\alias{ridge_reg}
\title{Perform a ridge regression}
\usage{
ridge_reg(formula, lambda, data)
}
\arguments{
\item{formula}{a formula}

\item{lambda}{a value for lambda}

\item{data}{a data.frame}
}
\value{
A ridge_reg object
}
\description{
This function perform ridge regression.
}
\examples{
fit <- ridge_reg(Sepal.Length ~.,1, iris)
fit
}
