% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_reg.r
\name{ridge_reg}
\alias{ridge_reg}
\title{Fit a ridge regression model}
\usage{
ridge_reg(formula, data, lambda)
}
\arguments{
\item{formula}{a formula}

\item{data}{a data.frame}

\item{lambda}{a numeric parameter}
}
\value{
An ridge_reg object
}
\description{
This function performs ridge regression.
}
\examples{
fit <- ridge_reg(Sepal.Length ~., 1.2, iris)
}
