% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_reg.r
\name{ridge_reg}
\alias{ridge_reg}
\title{Fit a Ridge Regression Model}
\usage{
ridge_reg(formula, lambda, data)
}
\arguments{
\item{formula}{a formula}

\item{lambda}{a penalty value}

\item{data}{a data.frame}
}
\value{
a ridge_reg object
}
\description{
This function performs ridge regression function.
}
\examples{
fit <- ridge_reg(Sepal.Length ~., lambda = 1.2121, iris)
summary(fit)
}
