% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge-model.R
\name{ridge_reg}
\alias{ridge_reg}
\title{Fit a ridge regression model}
\usage{
ridge_reg(formula, lambda, data)
}
\arguments{
\item{formula}{a formula}

\item{lambda}{a tuning parameter}

\item{data}{a data.frame}
}
\value{
An ridge object
}
\description{
This function passes parameters to the ridge regression function.
}
