% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_function.R
\name{ridge_reg}
\alias{ridge_reg}
\title{Fit a ridge regression model}
\usage{
ridge_reg(form, lambda, d)
}
\arguments{
\item{form}{a formula}

\item{lambda}{a lambda value}

\item{d}{a data.frame}
}
\value{
An ridge_reg object
}
\description{
This function passes parameters to the ridge_reg function.
}
\examples{
fit = ridge_reg(Sepal.Length~., 2, iris)
}
