% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridge_reg}
\alias{ridge_reg}
\title{Fit a ridge regression model}
\usage{
ridge_reg(form, lambda, d)
}
\arguments{
\item{lambda}{a vector}

\item{formula}{a formula}

\item{data}{a data.frame}
}
\value{
An list
}
\description{
This function runs ridge regression on dataset
}
