% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridge_reg}
\alias{ridge_reg}
\title{Fit a Ridge Regression Model}
\usage{
ridge_reg(form, lambda, d)
}
\arguments{
\item{form}{A formula}

\item{lambda}{regularization parameter}

\item{d}{A data.frame}
}
\value{
An list contain coefficients, lambda and formula
}
\description{
This function passes parameters to the ridge_reg function.
}
