% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge-reg.R
\name{ridge_reg}
\alias{ridge_reg}
\title{Fit a Ridge Regression model}
\usage{
ridge_reg(form, lambda, data)
}
\arguments{
\item{form}{a formula}

\item{lambda}{a penalty parameter}

\item{data}{a data.frame}
}
\value{
An ridge_reg object
}
\description{
This function passes parameters to the ridge regression function.
}
\examples{
fit <- ridge_reg(Sepal.Length ~.,1, iris)
summary(fit)
}
