% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge-regression.R
\name{ridge_regression}
\alias{ridge_regression}
\title{Ridge Regression Considering Collinearity}
\usage{
ridge_regression(form, data, lambda)
}
\arguments{
\item{form}{a formula with the legal format.}

\item{data}{a dataframe provided by the user.}

\item{lambda}{a penalty parameter specified by the user.}
}
\value{
a list including the coefficient estimates.
}
\description{
This is a ridge regression function taking into account collinear or nearly collinear regression variables.
}
\examples{
data(iris)
fit <- ridge_regression(Sepal.Length ~ ., iris)
fit$coefficients
}
