% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridge_regression}
\alias{ridge_regression}
\title{Ridge Regression}
\usage{
ridge_regression(formula, data, lambda = 0, contrasts = NULL)
}
\arguments{
\item{formula}{formula "y ~ ."}

\item{data}{dataframe}

\item{lambda}{ridge penalty term; default = 0}

\item{contrasts}{a list of contrasts}
}
\value{
List of beta coefficients
}
\description{
an implementation of ridge regression using singular value decomposition
}
\details{
This code was adapted from "A Computational Approach to Statistical Learning", Professor Kane's lecture materials, with the aid of Will Duke, and a collaoration with Christina Harden.
}
