% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridge_regression}
\alias{ridge_regression}
\title{Implement a ridge regression taking collinear features into account.}
\usage{
ridge_regression(formula, data, lambda, intercept = TRUE)
}
\arguments{
\item{formula}{An object of class formula to describe how the model is fitted.}

\item{data}{A data frame.}

\item{lambda}{An object of class numeric for the shrinkage parameter lambda.}

\item{intercept}{A logical value to indicate if intercept is included,
default = TRUE}
}
\value{
A list of vectors for coefficients, lambda and formula.
}
\description{
Implement a ridge regression taking collinear features into account.
}
\examples{
data(iris)
ridge_regression(formula = Sepal.Length ~ ., data = iris, lambda = 1.8)
}
