% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeregression.R
\name{ridge_regression}
\alias{ridge_regression}
\title{ridge regression}
\usage{
ridge_regression(form, data, lambda = 0)
}
\arguments{
\item{form}{A regression formula you want to use.}

\item{data}{A dataset you want to do ridge regression with.}

\item{lambda}{A number you want to specify for lambda in ridge regression, default is lambda = 0.}
}
\value{
A list containing the coefficients fo ridge regression, regression type and so on.
}
\description{
This is a function called ridge_regression, which does ridge regression that takes into account colinear regreesion variables.
}
\examples{
ridge_regression(Sepal.Length~.,iris,lambda=0.01)
}
