% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{ridge_regression}
\alias{ridge_regression}
\title{Ridge Regression}
\usage{
ridge_regression(formula, data, lambda = 0)
}
\arguments{
\item{formula}{a formula for ridge regression}

\item{data}{the data you want to regress on}

\item{lambda}{ridge penalty term}
}
\description{
Using SVD for ridge regression
}
\examples{
ridge_regression(mpg ~ cyl + disp, mtcars)
}
