% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridge_regression}
\alias{ridge_regression}
\title{Implement ridge regression}
\usage{
ridge_regression(X, y, lambda = 0)
}
\arguments{
\item{X}{numeric data matrix of predictors in a ridge regression model}

\item{y}{the response vector}

\item{lambda}{a hyperparameter that controls the penalty term}
}
\value{
Components of the fitted ridge regression model
}
\description{
This function can conduct ridge regression and export coefficients
}
