% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridge_regression}
\alias{ridge_regression}
\title{Ridge Regression}
\usage{
ridge_regression(form, data, lambda = 0.01, contrasts = NULL)
}
\arguments{
\item{form}{Formula of y~x1+x2+...}

\item{data}{The data set from which one desires to build the linear model}

\item{lambda}{Ridge regression parameter}

\item{contrasts}{List of contrasts for factor (categorical) variables}
}
\value{
List of beta coefficients derived from a linear regression
}
\description{
This is an implementation of ridge regression that takes into account colinear
(or nearly colinear) regression variables.
}
