% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridge_regression}
\alias{ridge_regression}
\title{Ridge regression function}
\usage{
ridge_regression(form, dataframe, lambda = 0)
}
\arguments{
\item{form}{ridge regression formula}

\item{dataframe}{data frame}

\item{lambda}{optional; default as 0}
}
\description{
This ridge Refression function will take into account colinear (or nearly colinear)
regression variables and return the coefficients
}
\examples{
data(iris)
ridge_regression(Sepal.Length ~ ., iris[,-5], 0.01)
}
