% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridge_regression}
\alias{ridge_regression}
\title{Ridge Regression}
\usage{
ridge_regression(X, Y, lambda)
}
\arguments{
\item{X}{The input design matrix.}

\item{Y}{The input response vector.}

\item{lambda}{The input penalty parameter.}
}
\description{
This is a function to implement a ridge regression function taking into account colinear (or nearly colinear) regression variables.
}
\examples{
model <- ridge_regression(X, Y, lambda=0.5)
model$coefficients
}
