% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridge_regression}
\alias{ridge_regression}
\title{Ridge Regression}
\usage{
ridge_regression(X, Y, lambda)
}
\arguments{
\item{X}{The input design matrix.}

\item{Y}{The input response vector.}

\item{lambda}{The input penalty parameter.}
}
\value{
A list of estimated coefficients.
}
\description{
This is a function to implement a ridge regression function taking into account colinear (or nearly colinear) regression variables.
}
\examples{
n <- 1000
p <- 5
beta <- c(1,-1,0,0,2)
set.seed(2)
X <- matrix(rnorm(n*p), nrow=n, ncol = p)
alpha <- 0.05
X[,1] <- X[,1]*alpha+X[,2]*(1 - alpha)
set.seed(1)
Y <- X \%*\% beta + rnorm(n)
model <- ridge_regression(X, Y, lambda=0.5)
model$coefficients
}
