% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridge_regression}
\alias{ridge_regression}
\title{Naive implementation of ridge regression.}
\usage{
ridge_regression(form, data, lambda = 0, contrasts = NULL)
}
\arguments{
\item{form}{a formula specifying the regression model.}

\item{data}{the data frame on which to fit the regression model.}

\item{lambda}{ridge penalty term.}

\item{contrasts}{contrasts.}
}
\value{
A list of beta coefficients.
}
\description{
a naive implementation of ridge regression using singular value decomposition.
}
\examples{
data(iris)
ridge_regression(Sepal.Length  ~ ., iris, lambda=0.1)

}
