% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridge_regression}
\alias{ridge_regression}
\title{ridge_regression}
\usage{
ridge_regression(form, data, lambda = 0, contrasts = NULL)
}
\arguments{
\item{form}{A symbolic description of the model to be fitted.}

\item{data}{A data frame or list that contains the variables in the model.}

\item{lambda}{Penalty added to beta}

\item{contrasts}{a list of contrasts}
}
\value{
The function returns a list of coefficients.
}
\description{
ridge_regression is used to fit linear models. It can be used to address the problem of colinearity and overfitting.
}
\examples{
ridge_regression(Sepal.Length~ ., iris)

}
