% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridge_regression}
\alias{ridge_regression}
\title{Ridge Regression}
\usage{
ridge_regression(formula, data, lambda = 0)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted.}

\item{data}{an data frame containing the variables in the model.}

\item{lambda}{the hyperparameter of the ridge regression. Default value is 0.}
}
\value{
A matrix containing the regression coeffeicients with the regression formula as an attribute.
}
\description{
This function performs ridge regression.
}
\examples{
data(iris)
scaled_iris <- as.data.frame(scale(iris[, -5], TRUE, TRUE))
fit.ridge_regression <- ridge_regression(Sepal.Length ~ .,
scaled_iris, lambda = 0.5)

}
