% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridge_regression}
\alias{ridge_regression}
\title{Ridge Regression}
\usage{
ridge_regression(formula, data, lambda = 0)
}
\arguments{
\item{formula}{a formula of the form y ~ .}

\item{data}{a dataframe}

\item{lambda}{ridge penalty term; default = 0}
}
\value{
A list of beta coefficients.
}
\description{
Ridge Regression
}
