% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridge_regression}
\alias{ridge_regression}
\title{Ridge Regression}
\usage{
ridge_regression(formula, data, lambda = 0, contrasts = NULL)
}
\arguments{
\item{formula}{a formula of the form y ~ .}

\item{data}{a dataframe}

\item{lambda}{ridge penalty term; default = 0}

\item{contrasts}{an optional list of contrasts}
}
\value{
A list of beta coefficients.
}
\description{
a minimal implementation of ridge regression employing singular value decomposition.
}
\details{
The bulk of the code is adapted from Professor Kane's book "A Computational Approach to Statistical Learning" as well as code examples from class. I added scaling and centering (modeled off of MASS:lm.ridge), reformatted the code to work in a function, edited output to pass tests, and added documentation.
}
