% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridge_regression}
\alias{ridge_regression}
\title{Ridge Regression}
\usage{
ridge_regression(form, d, contrasts = NULL, lambda = 0.1)
}
\arguments{
\item{form}{a formula}

\item{d}{a dataframe}

\item{contrasts}{contrasts argument}

\item{lambda}{ridge penalty term}
}
\value{
a list of beta coefficients
}
\description{
Implement a ridge regression function taking into account colinear (or nearly colinear) regression variables.
}
\examples{
\dontrun{
ridge_regression(body_mass_g ~ bill_depth_mm, penguins)
}
}
\author{
Yijun Yang
}
