% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q2_ridge_regression.R
\name{ridge_regression}
\alias{ridge_regression}
\title{conduct the ridge regression}
\usage{
ridge_regression(formula, data, lambda)
}
\arguments{
\item{formula}{The equation of the ridge regression model}

\item{data}{The data needed to be fitted by ridge model}

\item{lambda}{the parameter for the ridge regression}
}
\value{
A list of the coefficients of the ridge model
}
\description{
This function is aimed to conduct the ridge regression
}
\examples{
ridge_regression(formula = Sepal.Length ~ ., data = iris,lambda=1)
lm.ridge(formula = Sepal.Length ~ ., data = iris,lambda=1)
}
