% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridge_regression}
\alias{ridge_regression}
\title{ridge_regression}
\usage{
ridge_regression(form, dat, lambda = 0, contrasts = NULL)
}
\arguments{
\item{form}{A formula with the format of "Y ~ .".}

\item{dat}{A dataframe.}

\item{lambda}{The ridge penalty term lambda set default as 0.}

\item{contrasts}{A list of contrasts.}
}
\value{
A list of coefficients.
}
\description{
Compute a linear regression with gradient descent method.
}
