% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ridge_test.R
\docType{data}
\name{ridge_test}
\alias{ridge_test}
\title{Test data set for the second homework}
\description{
In homework 2 we will build ridge regression model.
This data set will be used to test the model.
}
