% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ridge-test.r
\docType{data}
\name{ridge_test}
\alias{ridge_test}
\title{A test set for the second homework}
\description{
In homework 2 we also look at the effect on the out-of-sample
mean square error as λ varies in ridge regressions. This data set gives us
a basis for exploration.
}
