% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ridge-test.r
\docType{data}
\name{ridge_test}
\alias{ridge_test}
\title{A dataset will be used for the test of ridge-regression}
\format{A data frame with 200 observations and 15 variables}
\source{
BIS557 HW2
}
\usage{
ridge_test
}
\description{
A dataset will be used for the test of ridge-regression
}
\keyword{datasets}
