% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ridge_test.R
\docType{data}
\name{ridge_test}
\alias{ridge_test}
\title{A data set for the ridge regression testing}
\description{
In homework 2 we explore the ridge regression.
This data set gives us a basis for exploration.
}
