\name{ridge_test}
\alias{ridge_test}
\docType{data}
\title{
Test Dataset for Ridge Regression
}
\description{
Test dataset for estimate coefficients by ridge regression.
}
\usage{data(ridge_test)}
\format{
  A data frame with 200 observations on the following 5 variables.
  \describe{
    \item{\code{x1}}{a numeric vector}
    \item{\code{x2}}{a numeric vector}
    \item{\code{x3}}{a numeric vector}
    \item{\code{x4}}{a numeric vector}
    \item{\code{y}}{a numeric vector; response variable}
  }
}
\keyword{datasets}
