% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.r
\docType{data}
\name{ridge_test}
\alias{ridge_test}
\title{A data set for the second homework}
\format{An object of class \code{data.frame} with 200 rows and 5 columns.}
\usage{
ridge_test
}
\description{
In homework 2 we explore ridge regressionon. This data set gives us a basis for #' exploration on the testing dataset.
}
\keyword{datasets}
