% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.r
\docType{data}
\name{ridge_test}
\alias{ridge_test}
\title{The testing data set for the second homework}
\format{An object of class \code{data.frame} with 200 rows and 5 columns.}
\usage{
ridge_test
}
\description{
A testing dataset used to test the ridge regression model in homework2.
}
\keyword{datasets}
