% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.r
\docType{data}
\name{ridge_test}
\alias{ridge_test}
\title{A test data for the second homework}
\description{
In homework 2 we explore the choice of lambda
on ridge regressions. This data set gives us a basis for exploration.
}
