% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.r
\docType{data}
\name{ridge_test}
\alias{ridge_test}
\title{A testing data set for the second homework}
\format{An object of class \code{data.frame} with 200 rows and 5 columns.}
\usage{
ridge_test
}
\description{
This is the testing dataset used to evaluate out of sample mse of the ridge function
}
\keyword{datasets}
