% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ridge_train.R
\docType{data}
\name{ridge_train}
\alias{ridge_train}
\title{Train data set for the second homework}
\description{
In homework 2 we will build ridge regression model.
This data set will help us train the model.
}
