% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ridge-train.r
\docType{data}
\name{ridge_train}
\alias{ridge_train}
\title{A training set for the second homework}
\description{
In homework 2 we explore the effect of numerical instability
on ridge regressions. This data set gives us a basis for exploration.
}
