% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ridge-train.r
\docType{data}
\name{ridge_train}
\alias{ridge_train}
\title{A dataset will be used for the train of ridge-regression}
\format{A data frame with 200 observations and 15 variables}
\source{
BIS557 HW2
}
\usage{
ridge_train
}
\description{
A dataset will be used for the train of ridge-regression
}
\keyword{datasets}
