% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.r
\docType{data}
\name{ridge_train}
\alias{ridge_train}
\title{A data set for the second homework}
\format{An object of class \code{data.frame} with 200 rows and 5 columns.}
\usage{
ridge_train
}
\description{
In homework 2 we explore ridge regression, especially the effect of lambda values
on ridge regressions. This data set gives us a basis for exploration on train dataset.
}
\keyword{datasets}
