% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ridge_train.R
\docType{data}
\name{ridge_train}
\alias{ridge_train}
\title{A data set for the ridge regression training}
\description{
In homework 2 we explore the ridge regression.
This data set gives us a basis for exploration.
}
