% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.r
\docType{data}
\name{ridge_train}
\alias{ridge_train}
\title{A training data set for the second homework}
\format{An object of class \code{data.frame} with 200 rows and 5 columns.}
\usage{
ridge_train
}
\description{
A training dataset for ridge regression in homework2.
}
\keyword{datasets}
