% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc-ridge_train.R
\docType{data}
\name{ridge_train}
\alias{ridge_train}
\title{The training data set for the second homework}
\format{An object of class \code{data.frame} with 200 rows and 5 columns.}
\usage{
ridge_train
}
\description{
A training dataset used to train the ridge regression model in homework2.
}
\keyword{datasets}
