% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_train.R
\docType{data}
\name{ridge_train}
\alias{ridge_train}
\title{A data set for the second homework}
\description{
In homework 2 we explore the effect of lambda values
on ridge regressions. This data set gives us a basis for exploration.
}
