% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-matrix.R
\docType{class}
\name{sparse.matrix-class}
\alias{sparse.matrix-class}
\title{Create a new class and function both called "sparse.matrix", and define addition, 
multiplication, and transpose of it.}
\description{
I set a new class called "sparse.matrix".
}
\section{Slots}{

\describe{
\item{\code{i}}{a vector indicating row number with non-zero elements}

\item{\code{j}}{a vector indicating columnn number with non-zero elements}

\item{\code{x}}{a vector indicating the value of each non-zero element}

\item{\code{dims}}{a vector indicating dimensions of the matrix}
}}

