% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-matrix.r
\docType{methods}
\name{sparse.matrix}
\alias{sparse.matrix}
\alias{+,sparse.matrix,sparse.matrix-method}
\alias{+,}
\alias{,sparse.matrix-method}
\alias{t,sparse.matrix-method}
\alias{t,}
\alias{sparse.matrix,ANY-method}
\alias{\%*\%,sparse.matrix,sparse.matrix-method}
\alias{\%*\%,}
\alias{sparse.matrix,sparse.matrix-method}
\title{Homework 4  - Question 3   - Reference to the code structure of setting classes by Justina Xie}
\usage{
\S4method{+}{sparse.matrix,sparse.matrix}(e1, e2)

\S4method{t}{sparse.matrix}(x)

\S4method{\%*\%}{sparse.matrix,sparse.matrix}(x, y)
}
\arguments{
\item{x}{value of the matrix}

\item{y, e1, e2}{a \code{sparse.matrix} object}
}
\value{
a the matrix of the operation: addition, multiplication and transpose
}
\description{
Homework 4  - Question 3   - Reference to the code structure of setting classes by Justina Xie

Addition

Transpose

Multiplication  - I completely dont understand this part, and this code is refernce to the Justina Xie, 
A small modification has been added on to the code.
}
\section{Slots}{

\describe{
\item{\code{dims}}{the dimensions of the matrix}
}}

