% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sparse_Matrix.R
\docType{methods}
\name{sparse.matrix}
\alias{sparse.matrix}
\alias{t,sparse.matrix-method}
\alias{t,}
\alias{sparse.matrix,ANY-method}
\alias{\%*\%,sparse.matrix,sparse.matrix-method}
\alias{\%*\%,}
\alias{sparse.matrix,sparse.matrix-method}
\alias{+,sparse.matrix,sparse.matrix-method}
\alias{+,}
\alias{sparse.matrix,}
\alias{sparse.matrix-method}
\title{S4 Class: sparse.matrix}
\usage{
\S4method{t}{sparse.matrix}(x)

\S4method{\%*\%}{sparse.matrix,sparse.matrix}(x, y)

\S4method{+}{sparse.matrix,sparse.matrix}(e1,e2)
}
\arguments{
\item{x}{value}

\item{y, e1, e2}{a \code{sparse.matrix} object}
}
\value{
a sparse.matrix object
}
\description{
S4 Class: sparse.matrix

Transpose method for sparse matrix

Multiplication method for sparse matrix

Addition method for sparse matrix
}
\section{Slots}{

\describe{
\item{\code{dims}}{dimension}
}}

