% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-matrix.r
\name{sparse.matrix}
\alias{sparse.matrix}
\title{create the class 'sparse.matrix'}
\usage{
sparse.matrix(i, j, x, dims)
}
\arguments{
\item{i}{row indices of elements in a sparse matrix}

\item{j}{col indices of elements in a sparse matrix}

\item{x}{value of the sparse matrix corresponding to i, j}

\item{dims}{dimensions of the sparse matrix}
}
\value{
A sparse.matrix object
}
\description{
create sparse matrix class using sparse.matrix
}
