% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_computation.r
\name{sparse.matrix}
\alias{sparse.matrix}
\title{Generate a sparse.matrix object from regular matrix}
\usage{
sparse.matrix(i, j, x, dims = c(max(i), max(j)))
}
\arguments{
\item{i}{the row coordinate of the entry}

\item{j}{the column coordinate of the entry}

\item{x}{the value at the entry}

\item{dims}{dimension of the matrix (if not specified, inferred from the largest i and j.)}
}
\value{
a 3 column (i:ith coordinate, j:jth coordinate, x:value) representation of the matrix
}
\description{
Generate a sparse.matrix object from regular matrix
}
