% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.matrix_class.R
\name{sparse.matrix}
\alias{sparse.matrix}
\title{Construct a sparse.matrix Object}
\usage{
sparse.matrix(a)
}
\arguments{
\item{a}{The object to make into a sparse.matrix}
}
\value{
The input object, but with "sparse.matrix" appended to the first spot in its class vector
}
\description{
This functions creates an instance of the sparse.matrix object, just adding "sparse.matrix" to the class vector. We assume that the object is a properly formatted data.frame.
}
\examples{
X <- sparse.matrix(X)
}
