% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-matrix.R
\name{sparse.matrix}
\alias{sparse.matrix}
\title{Sparse matrix and its operations}
\usage{
sparse.matrix(i, j, x, dims = c(max(i), max(j)))
}
\arguments{
\item{i}{row index of a non-zero element}

\item{j}{col index of a non-zero element}

\item{x}{value of the non-zero element}

\item{dims}{dimensions of the sparse matrix}
}
\value{
A sparse.matrix object
}
\description{
create a new class sparse.matrix consists of add, multiply and transpose
}
