% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.matrix.R
\name{sparse.matrix}
\alias{sparse.matrix}
\title{Use for creating "sparse.matrix" objects}
\usage{
sparse.matrix(i, j, x, dims = c(max(i), max(j)))
}
\arguments{
\item{i}{A vector indicating the row coordinates of non-zero entries of the matrix}

\item{j}{A vector indicating the column coordinate of non-zero entries of the matrix}

\item{x}{A vector indicating the values of entries at (i,j) positions of the matrix}

\item{dims}{The dimension of sparse matrix, default is the largest coordinates}
}
\value{
An "sparse.matrix" object
}
\description{
This function is used for creating "sparse.matrix" objects. 
"sparse.matrix" objects includes two parts. 
The first part is a sparse.matrix indicating the location and value of non-zero entries.
The second part is the dimension of matrix
}
\examples{
sm<- sparse.matrix(i = c(1, 2), j = c(1, 1), x = c(3, 1), dims = c(3, 2))
}
