% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework 4.R
\name{sparse.matrix}
\alias{sparse.matrix}
\title{Create a new class sparse.matrix}
\usage{
sparse.matrix(i, j, x, dims = c(max(i), max(j)))
}
\arguments{
\item{i}{A vector of row indices for non-zero entries}

\item{j}{A vector of column indices for non-zero entries}

\item{x}{A vector of the values for non-zero entries at the corresponding (i,j) positions}

\item{dims}{The dimension of sparse matrix, and default is the largest coordinates of i and j.}
}
\value{
A 'sparse.matrix' object.
}
\description{
This function is used for creating a new class called 'sparse.matrix'.
}
\examples{
sm3 <- sparse.matrix(i = rep(1, 3), j = 1:3, x = 1:3, dims = c(2, 3))
}
