% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_matrix.r
\name{sparse.matrix}
\alias{sparse.matrix}
\title{Sparse matrix and its operations}
\usage{
sparse.matrix(i, j, x, dims = c(max(i), max(j)))
}
\arguments{
\item{i}{row index of a nontrival element}

\item{j}{col index of a nontrival element}

\item{x}{value of the element corresponding to i, j}

\item{dims}{dimensions of the sparse matrix}
}
\value{
A sparse.matrix object
}
\description{
create sparse matrix using sparse.matrix, operations defined including add, multiply and transpose
}
