% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_matrix.r
\name{sparse.matrix}
\alias{sparse.matrix}
\title{Sparse matrix and operation functions}
\usage{
sparse.matrix(i, j, x, dims = c(max(i), max(j)))
}
\arguments{
\item{i:}{row index of non-zero elements}

\item{j:}{col index of non-zero elements}

\item{x:}{value of the elements corresponding to i, j}

\item{dims:}{dimension of the sparse matrix}
}
\value{
:  a sparse.matrix object
}
\description{
create a new class sparse.matrix that has add, multiply
             and transpose methods
}
