% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-matrix.R
\name{sparse.matrix}
\alias{sparse.matrix}
\title{Proble3: Sparse matrix's operations}
\usage{
sparse.matrix(i, j, x, dims = c(max(i), max(j)))
}
\arguments{
\item{i}{row index of a nontrival element}

\item{j}{col index of a nontrival element}

\item{x}{value of the element corresponding to i, j}

\item{dims}{dimensions of the sparse matrix
a <- sparse.matrix(i = c(1, 2), j = c(1, 3), x = c(3, 1))
b <- sparse.matrix(i = c(1, 2), j = c(3, 1), x = c(4.4, 1.2))}
}
\value{
A sparse.matrix object
}
\description{
create sparse matrix using sparse.matrix, operations defined including add, multiply and transpose
}
