% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-matrix.R
\name{sparse.matrix}
\alias{sparse.matrix}
\title{Create a new class "sparse.matrix"}
\usage{
sparse.matrix(i, j, x, dims = NULL)
}
\arguments{
\item{i}{a numeric vector for rows}

\item{j}{a numeric vector for columns}

\item{x}{a numeric vector specifying values in the matrix}

\item{dims}{a vector with dimensions of the sparse matrix}
}
\value{
A sparse matrix
}
\description{
This function Create a new class "sparse.matrix".
}
\examples{
sm0 <- sparse.matrix(i = c(1, 2), j = c(1, 1), x = c(1, 1))
}
