% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_matrix.R
\name{sparse.matrix}
\alias{sparse.matrix}
\title{Sparse matrix class}
\usage{
sparse.matrix(i, j, x, dims = c(max(i), max(j)))
}
\arguments{
\item{i}{row index of non-zero elements}

\item{j}{cololumn index of a non-zero elements}

\item{x}{numeric value of the element corresponding position i, j}

\item{dims}{dimensions of the sparse matrix}
}
\value{
A sparse.matrix object
}
\description{
creating sparse matrix class with addition, multiplication and transpose operations
}
