% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-matrix.R
\name{sparse_add}
\alias{sparse_add}
\title{Add sparse matrices}
\usage{
sparse_add(a, b)
}
\arguments{
\item{a}{a sparse.matrix object}

\item{b}{a sparse.matrix object}
}
\value{
a sparse.matrix object
}
\description{
This function adds two sparse matrices.
}
\examples{
a <- sparse.matrix(i = c(1, 2), j = c(1, 3), x = c(3, 1))
b <- sparse.matrix(i = c(1, 2), j = c(3, 1), x = c(4.4, 1.2))
sparse_add(a, b)
}
