% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_matrix.R
\name{sparse_multiply}
\alias{sparse_multiply}
\title{Multiply sparse matrices}
\usage{
sparse_multiply(a, b)
}
\arguments{
\item{a}{A list describing a sparse matrix.}

\item{b}{A list describing a sparse matrix.}
}
\value{
A  sparse.matrix object
}
\description{
Function is used to multiply two sparse matrices stored as triplet lists.
}
\examples{
a <- sparse.matrix(i = c(1, 2), j = c(1, 3), x = c(3, 1))
b <- sparse.matrix(i = c(1, 3), j = c(2, 1), x = c(4.4, 1.2))
sparse_multiply(a, b)
}
