% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-matrix.r
\name{t.sparse.matrix}
\alias{t.sparse.matrix}
\title{create the transpose function}
\usage{
\method{t}{sparse.matrix}(a)
}
\arguments{
\item{a}{a sparse.matrix object}
}
\value{
a sparse.matrix object
}
\description{
transpose operation defined in sparse matrix class
}
