% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.matrix_class.R
\name{t.sparse.matrix}
\alias{t.sparse.matrix}
\title{Transpose a sparse.matrix Object}
\usage{
t.sparse.matrix(a)
}
\arguments{
\item{a}{A sparse.matrix object}
}
\value{
A sparse.matrix object that is the matrix transpose of a
}
\description{
This method transposes a sparse.matrix object. Since we only record the non-zero entries in a sparse.matrix, all we need to do is reverse those entries' row and column indeces, as well as the dimensions.
}
\examples{
a <- sparse.matrix(i=c(1,2), j=c(1,1), x=c(3,1))
t(a)
}
