% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-matrix.R
\name{t.sparse.matrix}
\alias{t.sparse.matrix}
\title{Transpose}
\usage{
\method{t}{sparse.matrix}(a)
}
\arguments{
\item{a}{A list describing a sparse matrix.}
}
\description{
This funcdtion is used to transpose a sparse matrix
}
\examples{
a <- sparse.matrix(i = c(1, 2), j = c(1, 1), x = c(3, 1))
sparse_transpose(a)
}
