% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.matrix.R
\name{t.sparse.matrix}
\alias{t.sparse.matrix}
\title{Use for transposing a "sparse.matrix" object}
\usage{
\method{t}{sparse.matrix}(x)
}
\arguments{
\item{x}{A "sparse.matrix object"}
}
\value{
A "sparse.matrix" object
}
\description{
This function is used for transposing s "sparse.matrix" object
}
\examples{
sm0 <- sparse.matrix(i = c(1, 2), j = c(1, 1), x = c(1, 1))
t(sm0)
}
