% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-matrix.R
\name{t.sparse.matrix}
\alias{t.sparse.matrix}
\title{Transpose a sparse matrix}
\usage{
\method{t}{sparse.matrix}(a)
}
\arguments{
\item{a}{a sparse matrix}
}
\value{
A sparse matrix
}
\description{
This function transposes a sparse matrix.
}
\examples{
sm0 <- sparse.matrix(i = c(1, 2), j = c(1, 1), x = c(1, 1))
t(sm0)
}
